==> src/04_plot_outline.Rmd <==
---
title: "09_Paper_outline"
author: "Kolja Nenoff"
date: "2023-09-06"
output: html_document
---


#load



```{r}
rm(list = ls())
.libPaths("../../../R/x86_64-pc-linux-gnu-library/4.4/")
# devtools::install_github("BlakeRMills/MetBrewer")
libraries <- c("magrittr", "ggplot2", "data.table","patchwork","tidygraph","ggraph","igraph","tidyverse",
               "ggcorrplot","MetBrewer","dplyr","grid","gridExtra","ggpubr","cowplot","colorspace","RColorBrewer","gt","gtExtras",  "parallel","readxl","caret","Metrics","doMC","xgboost","energy","sf","RhpcBLASctl","rnaturalearth","rnaturalearthdata","ggpmisc","scales","SHAPforxgboost","countrycode")
for (lib in libraries) {
  if (!require(lib, character.only = TRUE)) {
    install.packages(lib)
    library(lib, character.only = TRUE) 
  }
}
```
# themes and setting

```{r}
#Theme settings
## plot theme
# color_seven<- ggokabeito::palette_okabe_ito()
col_pal<- brewer.pal(n = 8, name = "Dark2")
options(
  # set default colors in ggplot2 to colorblind-friendly
  # Okabe-Ito and Viridis palettes
  ggplot2.discrete.colour = "viridis",
  ggplot2.discrete.fill = "viridis",
  ggplot2.continuous.colour = "viridis",
  ggplot2.continuous.fill = "viridis",
  # set theme font and size
  book.base_family = "sans",
  book.base_size = 9
)



theme_set(
  theme_bw(
    base_size = getOption("book.base_size"),
    base_family = getOption("book.base_family")
  ) %+replace%
    theme(
      panel.grid.minor = element_blank(),
      legend.position = "bottom"
    )
)



```

# Cluster
```{r}
blas_set_num_threads(1)
omp_set_num_threads(1)
```
## Load preprocessed data

By running all script from the first 01 to the last script 04 the entire data will be stored in the DATA_DIR. However, to save time we recomannd to load the preprocessed data products stored in OSF. Link on github.
### set path
```{r}
#Cluster 

#Dir path
INPUT_DIR <- "../../input_data" #Preprocessed data from OSF
DATA_DIR <- "../data/04_data"
FIGURE_DIR <- "../figures/04_figures"


LIB_DIR<-"../libraries"
WDI_DIR <- file.path(INPUT_DIR,"231103_WDI") 
DIMRED_DIR<-INPUT_DIR
EXMORT_DIR <- file.path(INPUT_DIR,"230225_WHO_Excess_Mort/2022-03-25_covid-19_gem")
OWID_DIR<-file.path(INPUT_DIR,"231103_OWID_COVID")
OUTPUT_DIR<- "../output"

# File path
exmort_path <- file.path(EXMORT_DIR, "WHO_COVID_Excess_Deaths_EstimatesByCountry.xlsx")
dimred_path<-file.path(DIMRED_DIR,"01_wdi_data_cons_df_250.csv")
wdi_series_path <- file.path(WDI_DIR, "WDISeries.csv")
wdi_gf_path<-file.path(DIMRED_DIR, "01_wdi_gapfilled.csv")
toolbox_path <- file.path(LIB_DIR, "00_toolbox.R") # Source helper functions
owid_series_path<-file.path(OWID_DIR,"231103owid-covid-data.csv")
##Path to trained feature spaces (trained + Summarized 02)
raw_wdi_only_path<-file.path(OUTPUT_DIR,"../output/WDI_scaled_randomgrid_raw.RDS")
imputed_wdi_only_path<-file.path(OUTPUT_DIR,"WDI_randomgrid_scaled_imputed.RDS")
pc20_wdi_only_path<-file.path(OUTPUT_DIR,"WDI_randomgrid_scaled_pc20.RDS")
iso10_wdi_only_path<-file.path(OUTPUT_DIR,"WDI_randomgrid_scaled_iso10.RDS")
## Final model plus covariates (trained + Summarized 03)
pc20_model_path<-file.path(OUTPUT_DIR,"03_Xgboost_pca20_models")
iso10_model_path<-file.path(OUTPUT_DIR,"03_2410_Xgboost_E_iso10_models")
# iso5_model_path<-file.path(OUTPUT_DIR,"03_Xgboost_iso_models")
pathWDI_pc10<-file.path(INPUT_DIR,"03_data","WDI_randomgrid_scaled_pc10.RDS")
pathWDI_pc20<-file.path(INPUT_DIR,"03_data","WDI_randomgrid_scaled_pc20.RDS")
pathWDI_pc34<-file.path(INPUT_DIR,"03_data","WDI_scaled_pc34.RDS")
pathWDI_pc40<-file.path(INPUT_DIR,"03_data","WDI_scaled_pc40.RDS")
pathWDI_imputed<-file.path(INPUT_DIR,"03_data","WDI_randomgrid_scaled_imputed.RDS")
pathWDI_raw<-file.path(INPUT_DIR,"03_data","WDI_scaled_randomgrid_raw.RDS")
pathWDI_iso40<-file.path(INPUT_DIR,"03_data","WDI_randomgrid_scaled_iso40.RDS")
pathWDI_iso10<-file.path(INPUT_DIR,"03_data","WDI_randomgrid_scaled_iso10.RDS")
pathWDI_iso7<-file.path(INPUT_DIR,"03_data","WDI_randomgrid_scaled_iso7.RDS")
pathWDI_iso5<-file.path(INPUT_DIR,"03_data","WDI_randomgrid_scaled_iso5.RDS")
# Xgboost_E_iso10_models_path<-file.path(INPUT_DIR,"03_data","03_Xgboost_E_iso10_models")

pathWDI10_e_isomap<-file.path(INPUT_DIR,"03_data","WDI_randomgrid_scaled_ensemble_isomap10.RDS")
pathWDI20ppcaPCA<-file.path(INPUT_DIR,"03_data","WDI_randomgrid_scaled_ppca_pca20.RDS")
iso10_model_path<-file.path(INPUT_DIR,"03_data","03_2410_Xgboost_E_iso10_models")
Xgboostinput_complete_iso_path <-file.path(INPUT_DIR,"03_data","XboostDfEiso.RDS")
Covar_cases_path<-file.path(INPUT_DIR,"03_data","03_Xgboost_Covariates_noCases_final_withoutISO")
Newcasesmodel_path<-file.path(INPUT_DIR,"03_data","03_Xgboost_onlyVACSandNewcases_models")
wdi_data_cons_gf_path<-file.path(INPUT_DIR,"01_data/01_wdi_gapfilled.csv")
loading_path<-file.path(INPUT_DIR,"01_data/01_wdi_loadings.Rds")
RELATIVE_INDICATORS_FILE_NAME<-file.path(INPUT_DIR,"01_data/relative_indicators_2023_06.R")
covariates_path<-file.path(INPUT_DIR,"03_data","03_Xgboost_Covariates_final_withoutISO")
## Datasets for final analysis
pca20_final_df_path<-paste0("../data/03_data/XboostDfPCA.RDS")
iso10_final_df_path<-paste0("../data/03_data/XboostDfiso.RDS")
# Dimred contstruction data

```


```{r}
wdi_data_cons_gf<-fread(wdi_data_cons_gf_path)
loading_df<-readRDS(loading_path) %>% data.frame()

```

# Data prep



```{r}
## Excess Mort
ExcessMortbyCountry<-readxl::read_excel(exmort_path,sheet = 2)
excessCounrty_annual<-ExcessMortbyCountry %>% group_by(country,iso3,year) %>% 
  reframe(p_value_mort=sum(excess.mean)/sum(expected.mean)*100,ACM=sum(acm.mean))
excessCounrty_annual2 <- dplyr::mutate(excessCounrty_annual, identifier = paste0(iso3, year),"Country Code"=iso3) %>% select(-iso3)

## Dimred Data
dimred_250<-fread(dimred_path)


dimred_250$iso1 <- -dimred_250$iso1

# Flip the dimensions for other analysis
dimred_250$PC_normal_1 <- -dimred_250$PC_normal_1
dimred_250$PC_normal_5 <- -dimred_250$PC_normal_5
dimred_250$PC_normal_14 <- -dimred_250$PC_normal_14
dimred_250$PC_normal_6 <- -dimred_250$PC_normal_6
dimred_250$PC_normal_23 <- -dimred_250$PC_normal_23

## WDI
wdi_series <- fread(wdi_series_path) %>%
  setnames("Series Code", "Indicator Code")
##




formula="Y~X"

eck <- "+proj=eck4"

world <- ne_countries(scale = "medium", returnclass = "sf")  %>%
  st_wrap_dateline(options = c("WRAPDATELINE=YES","DATELINEOFFSET=180"), quiet = TRUE) %>%
  st_transform(crs = eck)

```

# 1. Background Figure Worldmap 1


```{r}
dimred2021_knn250<-dimred_250 %>% filter(!is.na(`Income Group`),`Income Group`!="",year%in%2020) %>% filter(!(`Country Code`%in% "PER")) %>% select(`Country Code`,`Income Group`) %>% mutate(incomegroup=case_when(grepl("High income",`Income Group`)~"HI",                                                                                                                  grepl("Upper middle income",`Income Group`)~"UMI",                                                                                                                         grepl("Lower middle income",`Income Group`)~"LMI",
                                                                       grepl("Low income",`Income Group`)~"LI"))
owid_data<-fread(owid_series_path)
owid_annual<-owid_data %>%  
  mutate(year=year(date))  %>% 
  filter(year%in%c(2020,2021)) %>% 
  data.table() %>% select(-date) %>% 
  filter(!is.na(continent)& continent!="") %>%   
  group_by(iso_code,location,year) %>% 
  reframe(
            Stringency_index=mean(stringency_index,na.rm=T),
            Stringency_index_var=var(stringency_index,na.rm=T),
            # Average_hosp_admissions=sum(weekly_hosp_admissions,na.rm=T)/population,
            # Average_ICU_admissions=sum(weekly_icu_admissions,na.rm=T)/population,
            Covid_Cases_per_million=sum(new_cases_per_million,na.rm=T),
            Owid_death_per_million =sum(new_deaths_per_million,na.rm=T),
            Owid_Exmort =mean(excess_mortality,na.rm=T),
            Vaccinations=sum(total_vaccinations_per_hundred,na.rm=T),
            Available_Hospital_beds=mean(hospital_beds_per_thousand,na.rm=T),
            Median_age=mean(median_age,na.rm=T),
            aged_65_older=mean(aged_65_older,na.rm=T),
            human_development_index=mean(human_development_index,na.rm=T),
            handwashing_facilities=mean(handwashing_facilities,na.rm=T),
            population_density=mean(population_density,na.rm=T),
            diabetes_prevalence=mean(diabetes_prevalence,na.rm=T),
            cardiovasc_death_rate=mean(cardiovasc_death_rate,na.rm=T)
            
            )%>%  arrange(iso_code,year) %>%  unique %>% mutate(identifier=paste0(iso_code,year))


owid_who<-inner_join(excessCounrty_annual,owid_annual,by=c("iso3"="iso_code","year"="year"))
dimred2021_knn250$incomegroup <- factor(dimred2021_knn250$incomegroup, levels=c("LI","LMI","UMI","HI"))
owid_who_worldbank<-inner_join(owid_who,dimred2021_knn250[,c("Country Code","Income Group","incomegroup")],by=c("iso3"="Country Code"))

owid_who_worldbank$`Income Group` <- factor(owid_who_worldbank$incomegroup, levels=c("LI","LMI","UMI","HI"))
owid_who_worldbank$incomegroup <- factor(owid_who_worldbank$incomegroup, levels=c("LI","LMI","UMI","HI"))

```





```{r}

formula <- y ~ x
default_color_palette_plotA <- scale_color_viridis_d(option="B",begin = 0,end = 0.9, name = "Income Group:  ",na.translate = F)
default_fill_palette <- scale_fill_viridis_d(option="B",begin = 0,end = 0.9, name = "Income Group:  ")


filtered_data <- owid_who_worldbank %>%
  filter(is.finite(Covid_Cases_per_million) & 
         is.finite(Owid_death_per_million) &
         Covid_Cases_per_million > 0 & 
         Owid_death_per_million > 0)

# Calculate the Pearson correlation coefficient and p-value
cor_test <- cor.test(log(filtered_data$Covid_Cases_per_million), 
                     log(filtered_data$Owid_death_per_million))

pearson_corr <- cor_test$estimate
pearson_pvalue <- cor_test$p.value
# Generate the plot
A <- owid_who_worldbank %>%
  ggplot(aes(x = log(Covid_Cases_per_million), y = log(Owid_death_per_million))) +
  geom_point(aes(color = `Income Group`), alpha = 0.9) +
  geom_smooth(method = "lm", formula = formula, se = FALSE, color = "red4") +
  theme(legend.position = "none", axis.line = element_line(linetype = "dotted")) +
  # Add Pearson correlation and p-value to the plot
   annotate("text", x = 7, y = Inf, label = sprintf('Pearson~r~"="~%.2f~~italic(p)~"< 0.001"', 
           pearson_corr, pearson_pvalue), 
           parse = TRUE, hjust = 1.1, vjust = 2, size = 3.5) +
  default_color_palette_plotA +  # Apply default color palette
  xlab("COVID-19 Cases (log)") +
  geom_hline(yintercept = 0, color = "grey30") +
  ylab("COVID-19 Death (log)")



# Remove rows where values for the variables are not finite (i.e., NA, Inf, or -Inf)
filtered_data_B <- owid_who_worldbank %>%
  filter(is.finite(Covid_Cases_per_million) & 
         is.finite(p_value_mort) &
         Covid_Cases_per_million > 0 & 
         p_value_mort > 0)

# Calculate the Pearson correlation coefficient and p-value
cor_test_B <- cor.test(log(filtered_data_B$Covid_Cases_per_million), 
                       log(filtered_data_B$p_value_mort))

pearson_corr_B <- cor_test_B$estimate
pearson_pvalue_B <- cor_test_B$p.value

# Generate the plot with Pearson correlation annotation
B <- filtered_data_B %>%
  ggplot(aes(x = log(Covid_Cases_per_million), y = log(p_value_mort))) +
  geom_point(aes(color = `Income Group`), alpha = 0.9) +
  geom_smooth(method = "lm", formula = formula, se = FALSE, color = "red4") +
  theme(legend.position = "none", axis.line = element_line(linetype = "dotted")) +
  # Annotate Pearson correlation and p-value
  annotate("text", x = 7, y = Inf, label = sprintf('Pearson~r~"="~%.2f~~italic(p)~"< 0.001"', 
           pearson_corr_B, pearson_pvalue_B), 
           parse = TRUE, hjust = 1.1, vjust = 2, size = 3.5) +
  default_color_palette_plotA +  # Apply default color palette
  xlab("COVID-19 Cases (log)") +
  geom_hline(yintercept = 0, color = "grey30") +
  ylab("Excess Mortality (log)")


C <- filtered_data  %>% 
  ggplot(aes(x = `Income Group`, y = Covid_Cases_per_million, fill = `Income Group`)) +  # Change color to fill
  geom_boxplot(color="black") +
  default_fill_palette +  # Adjust color palette
  default_color_palette_plotA+
  theme(legend.position = "none", axis.line = element_line(linetype = "dotted")) +
  ylab("COVID-19 Cases per million") +
  xlab("Income Group") +
  geom_jitter(aes(color = `Income Group`),alpha = 0.6) +scale_y_log10(breaks = trans_breaks("log10", function(x) 10^x),
              labels = trans_format("log10", math_format(10^.x))) +
  stat_kruskal_test(label = "p = {p.format}",label.y = (5.2))


# Plot D
D <- filtered_data %>%
  filter(!is.na(`Income Group`)) %>%
  ggplot(aes(x = `Income Group`, y = `p_value_mort`, fill = `Income Group`)) +  # Change color to fill
  geom_boxplot(color="black") +
  default_fill_palette+  # Adjust color palette
  default_color_palette_plotA+
  theme(legend.position = "asd", axis.line = element_line(linetype = "dotted")) +
 ylab("Excess Mortality in %") +
  xlab("Income Group") +
  geom_hline(yintercept = 0, color = "grey30") +
  geom_jitter(aes(color = `Income Group`),alpha = 0.6) +
  stat_kruskal_test(label = "p = {p.format}",label.y = 55)

legend <- 
  # create some space to the left of the legend
  owid_who_worldbank %>%
  filter(!is.na(`Income Group`)) %>%
  ggplot(aes(x = `Income Group`, y = `p_value_mort`, fill = `Income Group`)) +  # Change color to fill
  geom_boxplot(color="black") +
  default_fill_palette+  # Adjust color palette
  default_color_palette_plotA

legend1 = get_plot_component(legend, 'guide-box-bottom', return_all = TRUE)

# Combine plots
#Introduction_A <- ggarrange(A, B, C, D, legend = "none",labels = c("B","C","D","E"))
Introduction_A <- ggarrange( C, D, legend = "none",labels = c("b","c"))

Introduction_B = plot_grid(
  Introduction_A, 
  legend1, 
  ncol = 1,          # Arrange in one column (vertical stacking)
  rel_heights = c(1, 0.1)       # Align vertically
)

owid_who_worldbank %>%
  ggplot(aes(x = log(Owid_death_per_million), y = log(p_value_mort))) +
  geom_point(aes(color = `Income Group`), alpha = 0.9) +
  geom_smooth(method = "lm", formula = formula, se = FALSE, color = "red4") +
  theme(legend.position = "none", axis.line = element_line(linetype = "dotted")) +
  stat_fit_glance(
    method = "lm",
    label.y = "top",
    method.args = list(formula = formula),
    mapping = aes(
      label = sprintf(
        'R^2~"="~%.2f~~italic(p)~"="~%s',
        after_stat(r.squared), after_stat(format(round(p.value,digits = 3)))
      )
    ),
    parse = TRUE,
    size = 3
  ) +
  default_color_palette_plotA +  # Apply default color palette

  xlab("COVID-19 Death (log)") +
  geom_hline(yintercept = 0, color = "grey30") +
  ylab("Excess Mortality (log)") 


owid_data<-fread(owid_series_path)
owidMonth<-owid_data %>% group_by(iso_code) %>% 
  mutate(yearmon=paste0(year(date),"_",month(date))) %>% 
  group_by(iso_code,yearmon) %>% 
  reframe(
       Covid_Cases_per_million=sum(total_cases_per_million,na.rm=T),
       Stringency_index=mean(stringency_index,na.rm=T),
            Owid_Total_death =sum(total_deaths_per_million,na.rm=T),
            Owid_Exmort =mean(excess_mortality,na.rm=T))
ExcessMortMonth<-ExcessMortbyCountry %>% group_by(iso3) %>% 
  mutate(yearmon=paste0(year,"_",month)) %>% 
  group_by(iso3,yearmon) %>% 
  reframe(p_value_mort=sum(excess.mean)/sum(expected.mean)*100)
  
ulven<-inner_join(owidMonth,ExcessMortMonth,by=c("iso_code"="iso3","yearmon"="yearmon"))
pula<-ulven %>% 
  group_by(iso_code) %>% 
  # filter(!is.nan(Covid_Cases_per_million),!is.nan(Owid_Exmort)) %>% 
  reframe(
           dcor_case_deathEx=dcor(Covid_Cases_per_million, p_value_mort, index = 1.0),
           p_value_mort=mean(p_value_mort)
          # dcor_case_death=dcor(Covid_Cases_per_million, Owid_Exmort, index = 1.0),
          # dcor_death_Ex=dcor(Owid_Exmort, p_value_mort, index = 1.0)
          ) %>% arrange(desc(dcor_case_deathEx)) %>% 
  mutate(quartile = ntile(p_value_mort, 5)) %>% filter(!is.na(quartile))
  
kaluga<-pula %>% mutate(Quartiles=case_when(
  quartile%in%1~paste0(round(min(pula[pula$quartile%in%1,"p_value_mort"]))," - ",round(max(pula[pula$quartile%in%1,"p_value_mort"]))),
  quartile%in%2~paste0(round(min(pula[pula$quartile%in%2,"p_value_mort"]))," - ",round(max(pula[pula$quartile%in%2,"p_value_mort"]))),
  quartile%in%3~paste0(round(min(pula[pula$quartile%in%3,"p_value_mort"]))," - ",round(max(pula[pula$quartile%in%3,"p_value_mort"]))),
  quartile%in%4~paste0(round(min(pula[pula$quartile%in%4,"p_value_mort"]))," - ",round(max(pula[pula$quartile%in%4,"p_value_mort"]))),
  quartile%in%5~paste0(round(min(pula[pula$quartile%in%5,"p_value_mort"]))," - ",round(max(pula[pula$quartile%in%5,"p_value_mort"]))),
         ))



world_data <- merge(world,kaluga , by.x = "iso_a3_eh", by.y = "iso_code", all.x = TRUE) 


cap_value <- quantile(world_data$p_value_mort, 0.95,na.rm = T)  # Get the 95th percentile

world_data$p_value_mort_capped <- pmin(world_data$p_value_mort, cap_value)



cap_value <- quantile(world_data$p_value_mort, 0.95,na.rm = T)  # Get the 95th percentile

world_data$p_value_mort_capped <- pmin(world_data$p_value_mort, cap_value)


worldmap_excessMort <- ggplot() +
  geom_sf(data = world_data, aes(fill = p_value_mort_capped), lwd = 0.1, size = 6) +
  scale_fill_viridis(begin = 0.0,end=1,
    option = "D",
    name = "Excess Mortality  ")

worldmap_excessMort <- ggplot() +
  geom_sf(data = world_data, aes(fill = p_value_mort_capped), lwd = 0.1, size = 6) +
  scale_fill_gradient2(
                       name = "Excess Mortality (%):     ", 
                       low = "blue4",mid ="#FEFCC7" ,high = "#DD442F",
                       limits = c(-10, 30)) 



Part1<-ggarrange(worldmap_excessMort,ncol=1,labels = c("a",""),legend = "top",heights = 1.5,widths = 10)

Introduction_A <- ggarrange( C+ ylab("COVID-19 Cases / 1M"), D+ ylab("Excess Mortality (%)") , legend = "none",labels = c("b","c"))

Introduction_B = plot_grid(
  Introduction_A, 
  legend1, 
  ncol = 1,          # Arrange in one column (vertical stacking)
  rel_heights = c(1, 0.1)       # Align vertically
)
Introduction_B

cowplot::plot_grid(Part1,Introduction_B,ncol = 1,rel_widths =  c(1.2,0.9),rel_heights = c(1.2,0.8)) %>% 
   ggsave(file = "../figures/04_250718_introductionplot.png",device = "png",width = 5,height = 5.7)


```



# S1: Indicators of the pandemic health outcome Figure 1
```{r}
owid_data<-fread(owid_series_path)
owidMonth<-owid_data %>% group_by(iso_code) %>% 
  mutate(yearmon=paste0(year(date),"_",month(date))) %>% 
  group_by(iso_code,yearmon) %>% 
  reframe(
       Covid_Cases_per_million=sum(total_cases_per_million,na.rm=T),
       Stringency_index=mean(stringency_index,na.rm=T),
            Owid_Total_death =sum(total_deaths_per_million,na.rm=T),
            Owid_Exmort =mean(excess_mortality,na.rm=T))
ExcessMortMonth<-ExcessMortbyCountry %>% group_by(iso3) %>% 
  mutate(yearmon=paste0(year,"_",month)) %>% 
  group_by(iso3,yearmon) %>% 
  reframe(p_value_mort=sum(excess.mean)/sum(expected.mean)*100)
  
ulven<-inner_join(owidMonth,ExcessMortMonth,by=c("iso_code"="iso3","yearmon"="yearmon"))


Exmortestimationsandincome<-inner_join(x=ulven,y=dimred2021_knn250[,c("Country Code","incomegroup")],by=c("iso_code"="Country Code"))


pula<-ulven %>% 
  group_by(iso_code) %>% 
  filter(!is.nan(Covid_Cases_per_million),!is.nan(Owid_Exmort)) %>% 
  reframe(
          dcor_case_deathEx=dcor(Covid_Cases_per_million, p_value_mort, index = 1.0),
          dcor_case_death=dcor(Covid_Cases_per_million, Owid_Exmort, index = 1.0),
          dcor_death_Ex=dcor(Owid_Exmort, p_value_mort, index = 1.0),
          dcor_death_total=dcor(Owid_Exmort, p_value_mort, index = 1.0),
          Owid_Total_death
          ) %>% arrange(desc(dcor_case_deathEx))

# Merge your data frame with the world map data
world_data <- merge(world,pula , by.x = "iso_a3", by.y = "iso_code", all.x = TRUE)
# world_data <- merge(world,yo , by.x = "iso_a3", by.y = "iso_code", all.x = TRUE)
worldmap_supplement_1<-ggplot() +
  geom_sf(data = world_data, aes(fill = dcor_death_total) , lwd = 0.1,size=6) +
  theme_classic()+theme(legend.position = "not",)+scale_fill_viridis(option = "D")+ggtitle("COVID-19 Deaths (OWID) and \nCOVID-19 Cases (OWID)")
worldmap_supplement_2<-ggplot() +
  geom_sf(data = world_data, aes(fill = dcor_case_death) , lwd = 0.1,size=6) +
  theme_classic()+theme(legend.position = "not",)+scale_fill_viridis(option = "D")+ggtitle("Exces mortality (OWID) and \nCOVID-19 Cases (OWID)")
worldmap_supplement_3<-ggplot() +
  geom_sf(data = world_data, aes(fill = dcor_case_deathEx) , lwd = 0.1,size=6) +
  theme_classic()+theme(legend.position = "not",)+scale_fill_viridis(option = "D")+ggtitle("Excess mortality (WHO) and \nCOVID-19 Cases (OWID)")
worldmap_supplement_4<-ggplot() +
  geom_sf(data = world_data, aes(fill = dcor_death_Ex) , lwd = 0.1,size=6) +
  theme_classic()+theme(legend.position = "not",)+scale_fill_viridis(option = "D")+ggtitle("Excess mortality (OWID) and \nExmort (WHO) Correlation")
worldmap_supplement_supplement<-ggplot() +
  geom_sf(data = world_data, aes(fill = dcor_death_Ex) , lwd = 0.1,size=6) +
  theme_classic()+theme("dCor:",legend.position = "bottom",)+scale_fill_viridis(option = "D")+ggtitle("Excess mortality (OWID) and \nExcess mortality (WHO) Correlation")

legend1 = get_plot_component(worldmap_supplement_supplement, 'guide-box-bottom', return_all = TRUE)

plot_performance<-plot_grid(ggarrange(worldmap_supplement_1,worldmap_supplement_2,worldmap_supplement_3,worldmap_supplement_4,ncol=2,nrow = 2,heights = c(1,1),labels = c("a","b","c","d")), legend1,ncol = 1, rel_heights = c(1.3, 0.2))



plot_performance %>% 
 ggsave(file = "../figures/04_Introductions_world_maps_supplement.png",device = "png")

```



<!-- # Correlations -->
<!-- ```{r} -->
<!-- filtered_data <- owid_who_worldbank %>% -->
<!--   filter(is.finite(Covid_Cases_per_million) &  -->
<!--          is.finite(Owid_death_per_million) & -->
<!--          Covid_Cases_per_million > 0 &  -->
<!--          Owid_death_per_million > 0) -->

<!-- # Calculate the Pearson correlation coefficient and p-value -->
<!-- cor_test <- cor.test(log(filtered_data$Covid_Cases_per_million),  -->
<!--                      log(filtered_data$Owid_death_per_million)) -->

<!-- filtered_data <- owid_who_worldbank %>% -->
<!--   filter(is.finite(p_value_mort) &  -->
<!--          is.finite(Owid_death_per_million) & -->
<!--          Covid_Cases_per_million > 0 &  -->
<!--          Owid_death_per_million > 0) -->
<!-- filtered_data %>%  dim -->

<!-- cor_test_B <- cor.test(log(filtered_data_B$Owid_death_per_million),  -->
<!--                        log(filtered_data_B$p_value_mort)) -->

<!-- pearson_corr_B <- cor_test_B$estimate -->
<!-- pearson_corr_B -->
<!-- pearson_pvalue_B <- cor_test_B$p.value -->

<!-- owid_annual -->

<!-- ``` -->
<!-- ```{r} -->
<!-- owid_annual_extra<-owid_data %>%   -->
<!--   mutate(year=year(date))  %>%  -->
<!--   filter(year%in%c(2020,2021)) %>%  -->
<!--   data.table() %>% select(-date) %>%  -->
<!--   filter(!is.na(continent)& continent!="") %>%    -->
<!--   group_by(iso_code,location,year) %>%  -->
<!--   reframe( -->
<!--             Stringency_index=mean(stringency_index,na.rm=T), -->
<!--             Stringency_index_var=var(stringency_index,na.rm=T), -->
<!--             Average_hosp_admissions=sum(weekly_hosp_admissions,na.rm=T)/population, -->
<!--             Average_ICU_admissions=sum(weekly_icu_admissions,na.rm=T)/population, -->
<!--             Covid_Cases_per_million=sum(new_cases_per_million,na.rm=T), -->
<!--             tests_per_thousand=sum(new_tests_per_thousand,na.rm=T), -->

<!--             test_per_thousend=sum(new_tests_smoothed_per_thousand,na.rm = T), -->
<!--             Owid_death_per_million =sum(new_deaths_per_million,na.rm=T), -->
<!--             Owid_Exmort =mean(excess_mortality,na.rm=T), -->
<!--             Vaccinations=sum(total_vaccinations_per_hundred,na.rm=T), -->
<!--             Available_Hospital_beds=mean(hospital_beds_per_thousand,na.rm=T), -->
<!--             Median_age=mean(median_age,na.rm=T), -->
<!--             aged_65_older=mean(aged_65_older,na.rm=T), -->
<!--             human_development_index=mean(human_development_index,na.rm=T), -->
<!--             handwashing_facilities=mean(handwashing_facilities,na.rm=T), -->
<!--             population_density=mean(population_density,na.rm=T), -->
<!--             diabetes_prevalence=mean(diabetes_prevalence,na.rm=T), -->
<!--             cardiovasc_death_rate=mean(cardiovasc_death_rate,na.rm=T) -->

<!--             )%>%  arrange(iso_code,year) %>%  unique %>% mutate(identifier=paste0(iso_code,year)) -->
<!-- cor((owid_annual_extra$Covid_Cases_per_million),(owid_annual_extra$test_per_thousend),use = "complete.obs") -->
<!-- cor(owid_annual_extra$Covid_Cases_per_million,owid_annual_extra$Owid_death_per_million) -->
<!-- ``` -->


# S3: Properties of the Latent Space Representation of the WDI

```{r}
explainedvarofdimred<-readRDS(file.path(INPUT_DIR,"/01_data/01_dimred_qual.RDS"))
eisoexplained<-readRDS(file.path(INPUT_DIR,"01_data/eisomap.RDS"))

1-eisoexplained[1:6] 
explainedvarofdimred$`Ensemble Isomap`<-eisoexplained[1:35]
explainedvarofdimred$Dimension<-seq(nrow(explainedvarofdimred))
colnames(explainedvarofdimred)<-c("Isomap", "PCA"  ,"E-Isomap",  "Dimensions")
explainedvarofdimred$PCA<- explainedvarofdimred$PCA
dataExplained<-explainedvarofdimred %>% pivot_longer(names_to = "Method",cols = c("Isomap","PCA","E-Isomap")) #%>%
dataExplained$Method<-factor(dataExplained$Method,levels = c("PCA","Isomap","E-Isomap")) 
  
DimredPLotA<- dataExplained %>% ggplot( . ,aes(x=Dimensions,y=value,color=Method))+geom_point()+geom_line()+geom_hline(yintercept = 0.1,linetype="dotted")+ylab("Residual variance")+xlim(0,35)  + theme(
    legend.position = c(0.95, 0.95),
    legend.justification = c("right", "top"))+ylim(0,0.7)+scale_color_manual(values = c(col_pal[1],col_pal[2],"orange"))
dimred_250 %>%  filter(!is.na(PC_normal_1),!is.na(ISO_normal_1))-> subsetpcsiso



colors = MetPalettes
library(gridExtra)
a<-dimred_250 %>% ggplot()+geom_point(aes(x=ISO_normal_1,y=PC_normal_1),color=col_pal[3])+ theme_bw() + ylab("Principle component 1") +xlab("Isomap dimension 1")+scale_y_continuous(position = "right")
b<-ggplot(dimred_250) +geom_point(aes(x=ISO_normal_1,y=NY.GDP.PCAP.CD),color=col_pal[4])+ theme_bw() +ylab("GDP per capita")+ xlab("Isomap dimension 1")
c<-ggplot(dimred_250) +geom_point(aes(x=ISO_normal_1,y=SH.STA.MMRT),color=col_pal[5])+ theme_bw() +ylab("Maternal mortality ratio")+ xlab("Isomap dimension 1")+scale_y_continuous(position = "right")

cor(dimred_250$iso1,dimred_250$ISO_normal_1,use = "complete.obs")




d<-dimred_250 %>% ggplot()+geom_point(aes(y=NY.GDP.PCAP.CD,x=PC_normal_1),color=col_pal[4])+ theme_bw() + xlab("Principle Component 1") +ylab("GDP per capita")+scale_y_continuous(position = "right")

ElementR<-cowplot::plot_grid(DimredPLotA,a,b,c,cols = 2 ,align = "h",
          labels = c("a", "b", "c","d"))


ElementR %>%
  ggsave(file="../figures/04_figures/04_Dimredcompare_v2.png",device = "png",width = 7,height = 6)
```

# S4: Benchmark performance of different feature sets





```{r,raw}
wdi_rawData<-readRDS(pathWDI_raw)

wdi_imputedData<-readRDS(pathWDI_imputed)

wdi_pc40Data<-readRDS(pathWDI_pc40)
wdi_pc34Data<-readRDS(pathWDI_pc34)
wdi_pc20Data<-readRDS(pathWDI_pc20)
wdi_pc10Data<-readRDS(pathWDI_pc10)
wdi_iso40Data<-readRDS(pathWDI_iso40)
# wdi_iso20Data<-readRDS(pathWDI_iso20)
wdi_iso10Data<-readRDS(pathWDI_iso10)
wdi_iso7Data<-readRDS(pathWDI_iso7)
wdi_iso5Data<-readRDS(pathWDI_iso5)
WDI10_e_isomap<-readRDS(pathWDI10_e_isomap)
WDI20ppcaPCA<-readRDS(pathWDI20ppcaPCA)
WDI10_e_isomap$Xgboost_summary$RMSE %>%  median
WDI10_e_isomap$Xgboost_summary$RMSE %>%  IQR

```


```{r,raw}
performance_df<-rbind(wdi_rawData$Xgboost_summary %>% mutate(ident="A: WDI (Baseline)"),
                      wdi_imputedData$Xgboost_summary%>% mutate(ident="A: WDI (imputed)"),
                      wdi_pc40Data$Xgboost_summary%>% mutate(ident="B: PCA 40"),
                      wdi_pc34Data$Xgboost_summary%>% mutate(ident="B: PCA 34"),
                      wdi_pc20Data$Xgboost_summary%>% mutate(ident="B: PCA 20"),
                      wdi_iso40Data$Xgboost_summary%>% mutate(ident="C: Isomap 40"),
                      wdi_iso10Data$Xgboost_summary%>% mutate(ident="C: Isomap 10"),
                      wdi_iso7Data$Xgboost_summary%>% mutate(ident="C: Isomap 10"),
                      wdi_iso5Data$Xgboost_summary%>% mutate(ident="C: Isomap 5"),
                      WDI10_e_isomap$Xgboost_summary %>% mutate(ident="D: e-Isomap 10"),
                      WDI20ppcaPCA$Xgboost_summary %>% mutate(ident="D: ppca PCA 20")
                      )


myvector<-c("grey30","slategrey","steelblue","#7079BB","#B8509F","#73C9BC", "#C7D92D","#4EBA7C","orange","skyblue")

RMSE_decision<-performance_df %>%  
  mutate(Class=case_when(  grepl("^B",ident)~"B: PCA",
                           grepl("^C",ident)~"C: Isomap", 
                           grepl("^A",ident)~"A: Baseline",TRUE~"D: Others"
                           )) %>% filter(!grepl("D:",ident)) %>% 
  ggplot(aes(x=RMSE),"black")+geom_density(aes(fill=ident),alpha=0.5)+geom_vline(xintercept = median(wdi_rawData$Xgboost_summary$RMSE))+scale_fill_manual(values =myvector )+scale_color_manual(values =myvector )+xlab("Root Mean Square Error") + facet_wrap(~Class,ncol = 1)+
    theme(strip.text.x = element_text(hjust = 0, margin=margin(l=0.1)))


R2_decision<-performance_df  %>%  
  mutate(Class=case_when(  grepl("^B",ident)~"B: PCA",
                           grepl("^C",ident)~"C: Isomap", 
                           grepl("^A",ident)~"A: Baseline",TRUE~"D: Others"
                           )) %>% filter(!grepl("D:",ident)) %>% 
  ggplot(aes(x=Rsquared),"black")+geom_density(aes(fill=ident),alpha=0.5)+geom_vline(xintercept = median(wdi_rawData$Xgboost_summary$Rsquared))+scale_fill_manual(values =myvector )+scale_color_manual(values =myvector )+xlab(expression("Explained Variance")) + facet_wrap(~Class,ncol = 1) +
    theme(strip.text.x = element_text(hjust = 0, margin=margin(l=0.1)))
legend <- get_legend(
  # create some space to the left of the legend
  RMSE_decision + theme(legend.box.margin = margin(0, 0, 0, 12),legend.text=element_text(size=10))+scale_fill_manual("",values =myvector) +scale_color_manual(values =myvector ))

prow <- plot_grid(
  RMSE_decision + theme(legend.position="none"),
  R2_decision + theme(legend.position="none")+ylab(""),
  labels = c("I.", "II."),
  ncol = 2
)
plot_performance<-plot_grid(prow, legend,ncol = 1, rel_heights = c(1.3, 0.2))
plot_performance

# ggsave(plot_performance,file = "../figures/04_figures/04_performanceplots.png",device = "png",width = 9,height = 5)

```

## Plot Supplement
```{r,raw}


performance_df$ident <- factor(performance_df$ident, levels=c(
"A: WDI (Baseline)","A: WDI (imputed)","B: PCA 40" ,"B: PCA 34","B: PCA 20","C: Isomap 40","C: Isomap 10","C: Isomap 5","D: e-Isomap 10","D: ppca PCA 20" ))

RMSE_decision<-performance_df %>%  
  mutate(Class=case_when(  grepl("^B",ident)~"B: PCA",
                           grepl("^C",ident)~"C: Isomap", 
                           grepl("^A",ident)~"A: Baseline",TRUE~"D: Others"
                           )) %>% 
  ggplot(aes(x=RMSE),"black")+geom_histogram(aes(fill=ident),alpha=0.9)+geom_vline(xintercept = median(wdi_rawData$Xgboost_summary$RMSE))+scale_fill_manual(values =myvector )+scale_color_manual(values =myvector )+xlab("Root Mean Square Error") + facet_wrap(~Class,ncol = 1)+
    theme(strip.text.x = element_text(hjust = 0, margin=margin(l=0.1)))


R2_decision<-performance_df  %>%  
  mutate(Class=case_when(  grepl("^B",ident)~"B: PCA",
                           grepl("^C",ident)~"C: Isomap", 
                           grepl("^A",ident)~"A: Baseline",TRUE~"D: Others"
                           )) %>% 
  ggplot(aes(x=Rsquared),"black")+geom_histogram(aes(fill=ident),alpha=0.9)+geom_vline(xintercept = median(wdi_rawData$Xgboost_summary$Rsquared))+scale_fill_manual(values =myvector )+scale_color_manual(values =myvector )+xlab(expression("Explained Variance")) + facet_wrap(~Class,ncol = 1) +
    theme(strip.text.x = element_text(hjust = 0, margin=margin(l=0.1)))

legend <- get_legend(
  # create some space to the left of the legend
  RMSE_decision + theme(legend.box.margin = margin(0, 0, 0, 12),legend.text=element_text(size=10))+scale_fill_manual("",values =myvector) +scale_color_manual(values =myvector ))

prow <- plot_grid(
  RMSE_decision + theme(legend.position="none"),
  R2_decision + theme(legend.position="none")+ylab(""),
  labels = c("I.", "II."),
  ncol = 2
)
plot_performance_supp<-plot_grid(prow, legend,ncol = 1, rel_heights = c(1.3, 0.2))
plot_performance_supp
# ggsave(plot_performance_supp,file = "../figures/04_figures/04_performanceplots_supplement.png",device = "png",width = 9,height = 5)

```



## Supplement Table 1 Predictivte performance

```{r}
wdi_rawData$Xgboost_summary$RMSE %>%  summary()
wdi_imputedData$Xgboost_summary$RMSE %>%  summary()
wdi_pc20Data$Xgboost_summary$RMSE %>%  summary()
wdi_iso10Data$Xgboost_summary$RMSE %>%  summary()
wdi_ppcaPCA20$Xgboost_summary$RMSE %>%  summary()
wdi_e_iso10Data$Xgboost_summary$RMSE %>%  summary()


wdi_rawData$Xgboost_summary$Rsquared %>%  summary()
wdi_imputedData$Xgboost_summary$Rsquared %>%  summary()
wdi_pc20Data$Xgboost_summary$Rsquared %>%  summary()
wdi_iso10Data$Xgboost_summary$Rsquared %>%  summary()
# wdi_ppcaPCA20$Xgboost_summary$Rsquared %>%  summary()
# wdi_e_iso10Data$Xgboost_summary$Rsquared %>%  summary()
```

# 3.1 Global predictive model of COVID-19 excess mortality Figure 3

```{r}

Xgboostinput_complete_iso <-readRDS(Xgboostinput_complete_iso_path)
Xgboostinput_iso<-Xgboostinput_complete_iso%>% rename("Country.Code"=`Country Code`,
                                                     "Diabetes (%)" =`Diabetetes (%)`,
                                                     "COVID-19 cases"=`Covid cases per million`,
                                                     "Comorbity"=`Comorbidity index` )
```


```{r}
# correlation of the variables
look<-combn(Xgboostinput_complete_iso[,2:ncol(Xgboostinput_complete_iso)] %>% colnames,2)
corvar<-data.frame()
for(i in 1:ncol(look)){
 try(ulf<-data.frame("var1"=look[1,i],"var2"=look[2,i],"cor"=as.numeric(unlist(cor(Xgboostinput_complete_iso[,look[1,i],with = F],Xgboostinput_complete_iso[,look[2,i],with = F],use = "complete.obs")))))
corvar<-rbind(corvar,ulf)
}

```




```{r}
colors <- viridisLite::viridis(n = 4, option = "D")

modeliso10<-readRDS(iso10_model_path)
importantFeatures<-modeliso10$Feature_summary %>%  group_by(feature) %>% 
   reframe(Shapley_mean=mean(abs(Shap_value_mean)),Importance=mean(importance)) %>%  arrange(desc(Shapley_mean))
modeliso10$Xgboost_summary$Rsquared %>%  max


groupedModeloutput<-modeliso10$Feature_summary %>% group_by(run) %>%
  mutate(COVID_VACS=case_when(grepl("Covid|Vac|Ox",feature)~"Pandemic \nindicators",
                              grepl("iso",feature)~"cNFC",
                              TRUE~"Epidem. \nfeatures"))

datablopA<-modeliso10$Feature_summary %>%  group_by(feature)%>%
  filter(feature%in%importantFeatures$feature[1:10]) %>%
  mutate(feature=case_when(grepl("Covid",feature)~"Covid-19 Cases",TRUE~feature),type=case_when(grepl("Covid|Vac|Ox",feature)~"Pandemic \nindicators",
                              grepl("iso",feature)~"cNFC",
                              TRUE~"Epidem. \nfeatures"))



datablop<-groupedModeloutput %>%  group_by(run,COVID_VACS)   %>% reframe(Shap_value_mean=sum(Shap_value_mean)) %>% mutate(feature=COVID_VACS)
interRes<-datablop %>% group_by(COVID_VACS) %>% reframe(aplha=sum(Shap_value_mean))

interRes$aplha/sum(interRes$aplha)

explainedVar<-datablop %>%  group_by(feature)   %>% reframe(Shap_value_mean1=mean(Shap_value_mean),
                                                            Shap_value_median1=median(Shap_value_mean),
                                                            Shap_value_var1=var(Shap_value_mean),
                                                            ) 
modelShap<-datablop %>%  group_by(run)   %>% reframe(Shap_value_mean1=sum(Shap_value_mean)  )


datablop %>% group_by(COVID_VACS) %>% reframe(median=paste("|SHAP|: Median =",round(median(Shap_value_mean),1)),value=paste("IQR = ",round(IQR(Shap_value_mean),1)))

SummedupPredictions<-datablop  %>% group_by(run) %>%  reframe(sum(Shap_value_mean
))


datablop$type<-factor(datablop$COVID_VACS,levels = c("Pandemic \nindicators","Epidem. \nfeatures" ,"cNFC"))
datablopA$typeA<-factor(datablopA$type,levels = c("Pandemic \nindicators","Epidem. \nfeatures" ,"cNFC"))

SectionOver<-rbind(datablop) %>% ggplot( . , aes(x=reorder(feature,(Shap_value_mean)), y=Shap_value_mean,fill=type))+
  geom_boxplot()  + ylab("Importance (SHAP)")+ xlab("")+ylim(0,11) +scale_fill_manual(values=colors) +theme_bw() +theme(legend.position = "bottom" )
SectionOver
```




```{r}
datablopA$feature<-str_replace(datablopA$feature,"iso","cNFC ")
datablopA$feature<-str_replace(datablopA$feature,"Vaccinations per hundred","Vaccinations")
SectionUnder<-datablopA %>% ggplot( . , aes(x=reorder(feature,(Shap_value_mean)), y=Shap_value_mean,fill=typeA))+
  geom_boxplot(alpha=0.6)  + ylab("Importance (SHAP)")+ xlab("")+scale_fill_manual(values=colors)+theme(legend.position = "F")+theme_bw()+theme(legend.position = "bottom") +
  scale_x_discrete(guide = guide_axis(n.dodge = 2))


```


```{r}




Element_one<-cowplot::plot_grid(SectionOver,SectionUnder,labels = c("A", "B"),rel_heights =   c(0.8,2.3),ncol = 1)


datablop<-groupedModeloutput %>%  group_by(run,feature)   %>% reframe(Shap_value_mean=sum(Shap_value_mean)) 
runsum<-datablop %>%  group_by(run)   %>% reframe(absolute=sum(Shap_value_mean)) 
runsum1<-merge(datablop,runsum,by="run") %>%  mutate(contributionperrun=Shap_value_mean/absolute)
discussion<-runsum1 %>%  group_by(feature)   %>% reframe(percentcontribtution=(median(contributionperrun)*100),iqr=(IQR(contributionperrun))*100) %>% arrange(desc(percentcontribtution))


```



```{r}

datablop<-groupedModeloutput %>%  group_by(run,COVID_VACS)   %>% reframe(Shap_value_mean=sum(Shap_value_mean)) %>% mutate(feature=COVID_VACS)
datablop$type<-factor(datablop$COVID_VACS,levels = c("Pandemic \nindicators","Epidem. \nfeatures" ,"cNFC"))
datablopA$typeA<-factor(datablopA$type,levels = c("Pandemic \nindicators","Epidem. \nfeatures" ,"cNFC"))

SectionOver<-rbind(datablop) %>% ggplot( . , aes(x=reorder(feature,(Shap_value_mean)), y=Shap_value_mean,fill=type))+
  geom_boxplot()  + ylab("Importance (SHAP)")+ xlab("")+ylim(0,11) +scale_fill_manual(values=colors) +theme_bw() +theme(legend.position = "as" ,
        panel.border = element_rect(colour = colors[4], fill=NA, linewidth=1) )+theme(legend.position = "bottom",
        panel.border = element_rect(colour = colors[4], fill=NA, linewidth=1,
        linetype = "solid"))+ xlab("Group contribution in integrated model")

datablopA$feature<-datablopA$feature %>% str_replace_all("^Vaccinations per hundred$","Vaccinations/100")
datablopA$feature<-datablopA$feature %>% str_replace_all("^Vaccinations","Vaccinations/100")
datablopA$feature<-datablopA$feature %>% str_replace_all("^Covid-19 Cases$","Cases/1M")
SectionUnder<-datablopA %>% ggplot( . , aes(x=reorder(feature,(Shap_value_mean)), y=Shap_value_mean,fill=typeA))+
  geom_boxplot(alpha=0.8)  + ylab("Importance (SHAP)")+ xlab("")+scale_fill_manual(values=colors)+theme(legend.position = "F")+theme_bw()+theme(legend.position = "bottom",
        panel.border = element_rect(colour = colors[4], fill=NA, linewidth=1,
        linetype = "solid")) +
  scale_x_discrete(guide = guide_axis(n.dodge = 2)) + xlab("Top 10 feature contributions in the integrated model")
SectionUnder
```

### plot
```{r}
R2_decision
ggarrange(ggarrange(R2_decision,SectionOver+theme(legend.position ="as"),labels = c("a", "b")),SectionUnder+theme(legend.position = "as"),ncol=1,labels = c("", "c")) # %>%  ggsave(file="../figures/04_figures/04_2505_condensed.png",device = "png",width = 9,height = 7)


# ggarrange(ggarrange(R2_decision,SectionOver+theme(legend.position ="as"),labels = c("a", "b")),SectionUnder+theme(legend.position = "as"),ncol=1,labels = c("", "c"))  %>%  ggsave(file="../figures/Lancet_submission/Results1.pdf",device = "pdf",width = 9,height = 7)
```





## Extra: Forceplot 


```{r}

require(SHAPforxgboost)
bestSEED<-modeliso10$Xgboost_summary %>% arrange(desc(Rsquared))
set.seed(bestSEED$SeedNum[1000])

Xgboostinput_complete_iso <-readRDS(Xgboostinput_complete_iso_path)
Xgboostinput_iso<-Xgboostinput_complete_iso%>% rename("Country.Code"=`Country Code`,
                                                     "Diabetes (%)" =`Diabetetes (%)`,
                                                     "COVID-19 cases"=`Covid cases per million`,
                                                     "Comorbity"=`Comorbidity index` )

dimred_250$incomegroup <- factor(dimred_250$`Income Group`, levels=c("Low income","Lower middle income","Upper middle income","High income"))
additionalVariables<-dimred_250 %>% filter(year%in%2019)%>%  select(`Country Code`,incomegroup,NY.GDP.PCAP.CD) %>% 
  mutate(incomegroup=case_when(
    incomegroup%in%"Low income"~"LI",
    incomegroup%in%"Lower middle income"~"LMI",
    incomegroup%in%"Upper middle income"~"UMI",
    incomegroup%in%"High income"~"HI",
    TRUE~incomegroup))%>%  unique
additionalVariables$incomegroup <- factor(additionalVariables$incomegroup, levels=c("LI","LMI","UMI","HI"))

Label<-"p_value_mort"
features<-Xgboostinput_iso %>% select(-`Country.Code`,-p_value_mort) %>%  colnames()
features_iso<-Xgboostinput_iso %>% select(-`Country.Code`,-p_value_mort) %>%  colnames()
features_iso<-features_iso[grep("cases|iso",features_iso)]
Xgboostinput_iso<-Xgboostinput_iso %>% filter(!((`Country.Code`%in%"PER")))
#features<-features_iso
features<-modeliso10$Feature_summary$feature %>%  unique
features<-  c("Vaccinations","COVID-19 cases","iso1","iso2","iso3","iso4","iso5","iso6","iso7","iso8","iso9","iso10","Obesity (%)","Diabetes (%)","Comorbity","Urban area (%)","Oxford Stringency Index" )
y_var <-  Label
dataX <- as.matrix(Xgboostinput_iso %>% select(features))

# hyperparameter tuning results
param_list <- list(objective = "reg:squarederror",  # For regression
  nrounds = c(100,50),  # Number of boosting rounds
  max_depth = c(3, 2),     # Maximum tree depth
  eta = c(0.05),    # Learning rate
  gamma = c(0, 5),       # L2 regularization
 colsample_bytree = c( 0.9,1),
  subsample = c(0.7, 0.8, 0.9),  # Subsample ratio
  min_child_weight = c(5, 10)  # Minimum child weight
                   )

mod <- xgboost::xgboost(data = dataX, 
                        label = as.matrix(Xgboostinput_iso[[y_var]]), 
                        params = param_list, nrounds = 1000,
                        verbose = FALSE, nthread = 30,
                        early_stopping_rounds = 100)

shap_values <- shap.values(xgb_model = mod, X_train = dataX)


shap_values$mean_shap_score %>%  sum()

shap_values2<-shap_values$shap_score

plot_data <- shap.prep.stack.data(shap_contrib = shap_values2, top_n = 16, n_groups = 1)



plot_iso<-cbind(plot_data %>% arrange(ID),a=Xgboostinput_iso$Country.Code,b=Xgboostinput_iso$iso1)
# 
muhara<-inner_join(plot_iso,additionalVariables,by=c("a"="Country Code"))
muhara2<-muhara  %>% mutate(`Pandemic Indicators`=`COVID-19 cases`+Vaccinations+`Oxford Stringency Index`,
                   `cNFC`=iso1+iso2+iso3+iso4+iso5+iso6+iso7+iso8+iso9,
                   `Epidemiological Features`=`Obesity (%)`+`Diabetes (%)`+`rest_variables`+`Comorbity`,
                   ) %>%  select(-c(iso1,iso2,iso3,iso4,iso5,iso6,iso7,iso8,iso9,`Obesity (%)`,`rest_variables`,`Comorbity`,`COVID-19 cases`,Vaccinations,`Oxford Stringency Index`,`Diabetes (%)`))


regionalised<-dimred_250 %>%  select(`Country Code`,RegionSoviet) %>%  unique
muhara2%>% 
  merge( . ,regionalised,by.x = "a",by.y ="Country Code") %>%  
  ggplot(aes(x=`Pandemic Indicators`,y=cNFC,color=RegionSoviet)) +geom_point()
    

muhara2      # %>% fwrite("2502_ForcePlotdata.csv")           


plot_iso2<-muhara2 %>%  arrange(incomegroup,NY.GDP.PCAP.CD)%>%  mutate(sorted_id=1:n()) 
plot_iso2 %>% select(-a,-b,-NY.GDP.PCAP.CD,-incomegroup)
aldender<-shap.plot.force_plot((plot_iso2 %>% select(-a,-b,-NY.GDP.PCAP.CD,-incomegroup)), zoom_in_location = 295,y_zoomin_limit = c(-20,20))+scale_fill_viridis_d(begin = 0,end=0.5,
    name = "Contribution",na.translate = F)+theme(legend.position = "None")

where<-plot_iso2 %>%  group_by(incomegroup) %>% summarise(regions=max(sorted_id))
options(book.base_family = "sans",
  book.base_size = 8)

isoforcePLot<-shap.plot.force_plot((plot_iso2 %>% select(-a,-b,-incomegroup,-NY.GDP.PCAP.CD,-`Urban area (%)`)), zoom_in_location = 285,y_zoomin_limit = c(-10,10))+scale_fill_viridis_d(begin = 0,end=0.5,
    name = "Contribution",na.translate = F)+theme(legend.position = "None")+ geom_vline(xintercept =where$regions,color="grey30" ) + annotate("text", x = where$regions, y = 37, label = where$incomegroup, vjust = -1, hjust = 1)+scale_y_continuous(position = "right")+ xlab("Countries ordered by GDP per Capita")+ylab("SHAP values by feature type")  +   theme_bw(
    base_size = getOption("book.base_size"),
    base_family = getOption("book.base_family")
  ) %+replace%
    theme(
      panel.grid.minor = element_blank(),
      legend.position = "bottom"
    )+scale_fill_manual(values = c(colors[c(1,3,2,5)]))

isoforcePLot




```

#### merge
```{r}

cool<-cowplot::plot_grid(Element_one,cowplot::plot_grid(R2_decision+theme(legend.position = "ad"),isoforcePLot,ncol = 1,rel_heights =   c(0.8,2.3),labels = c("", "D")),labels = c(" ", "C"),rel_heights =   c(0.8,1),ncol = 2)
cool
# cool<-cowplot::plot_grid(Element_one,cowplot::plot_grid(isoforcePLot,ncol = 1,rel_heights =   c(0.8,2.3),labels = c("", "D")),labels = c(" ", "C"),rel_heights =   c(0.8,1),ncol = 2)

legend1 = get_plot_component(R2_decision, 'guide-box-top', return_all = TRUE)
legend_centered = plot_grid(
  NULL, legend1, NULL, 
  ncol = 3, 
  rel_widths = c(1, 1, 1) # Adjust to center the legend
)

final_plot = plot_grid(
  cool, 
  legend1, 
  ncol = 1,          # Arrange in one column (vertical stacking)
  rel_heights = c(1, 0.08),  # Adjust heights, give more space to panelA
  align = "v"        # Align vertically
)
final_plot
#final_plot %>%  ggsave(file="../figures/04_figures/04_2505_isoResults.png",device = "png",width = 9,height = 10)
```

# S7 Final Model summary
```{r}
# Newcasesmodel<-readRDS(Newcasesmodel_path)$Xgboost_summary
# Covar_cases<-readRDS(Covar_cases_path)$Xgboost_summary
# pathWDI10_e_isomap<-file.path(INPUT_DIR,"03_data","03_2410_Xgboost_E_iso10_models")
# iso10<-readRDS(pathWDI10_e_isomap)$Xgboost_summary
CompleteModel<- modeliso10$Xgboost_summary
CompleteModel$Rsquared %>%  summary

# Newcasesmodel$type<-"Pandemic \nindicators"
# # iso10$type<-"cNFC"
# CompleteModel$type<-"Final Model"
# Covar_cases$type<-"Epidem. \nfeatures"
# these<-bind_rows(Newcasesmodel,iso10,Covar_cases,CompleteModel)
# these$type<-factor(these$type,levels = c("Pandemic \nindicators","Epidem. \nfeatures","cNFC","Final model"))
these<-CompleteModel

# summarymaker<-these %>% group_by(type) %>%
#   reframe(Rsquared_median=median(Rsquared),
#           Rsquared_IQR=IQR(Rsquared),
#           Rsquared_max=max(Rsquared),)
# paste0("$R^2$: Median=\num{",(round(summarymaker[4,2],3)*100),"}",
#        "IQR=\num{",(round(summarymaker[4,3],3)*100),"}"
#        )


# covariates<-readRDS(covariates_path)
# summary_covariates<-covariates$Feature_summary %>%  group_by(feature)   %>% reframe(Shap_value_mean1=mean(Shap_value_mean),
#                                                             Shap_value_median1=median(Shap_value_mean),
#                                                             Shap_value_var1=var(Shap_value_mean),
#                                                             ) %>%  arrange(desc(Shap_value_mean1))



### Sumamrise the performance


plot_df <- CompleteModel %>%
  pivot_longer(cols = c(Rsquared, RMSE), names_to = "Metric", values_to = "Value")




 colors <- viridisLite::viridis(n = 4, option = "F")

## Plot

# colors <- viridisLite::viridis(n = 4, option = "F")
options(scipen=999)
# kw_test <- kruskal.test(Rsquared ~ type, data = these)
# pval_text <- ifelse(kw_test$p.value < 0.001, 
#                     "Kruskal-Wallis: p < 0.001", 
#                     paste0("Kruskal-Wallis: p = ", signif(kw_test$p.value, 3)))
colors <- ggokabeito::palette_okabe_ito()[] 
# R2_decision<-these %>% 
#   ggplot(aes(x=Rsquared,fill=type))+geom_density(alpha=0.7)+xlab(expression("Explained Variance")) +scale_y_continuous(position = "right") +scale_fill_manual("Components:",values = colors)+theme_bw()+theme(legend.position = "asd")
# these<-bind_rows(Newcasesmodel,iso10,Covar_cases,CompleteModel)
# these[these$type=="Final Model","type"]<-"Integrated \nmodel"
# these$type<-factor(these$type,levels = c("Pandemic \nindicators","Epidem. \nfeatures","cNFC","Integrated \nmodel"))

R2_decision <- these%>% 
  ggplot(aes(x = Rsquared)) +
  geom_histogram(position = "identity", alpha = 0.8, bins = 40, fill = "#0191B4") +
  xlab(expression("Explained variance (R"^2*")")) +
  scale_y_continuous(position = "right") +
  theme_bw() +
  theme(
    legend.position = "left",
    panel.grid.minor = element_blank()
  ) +
  ylab("") 

RMSE_decision <- these %>% 
  ggplot(aes(x = RMSE))+
  geom_histogram(position = "identity", alpha = 0.8, bins = 40, fill = "#E7B7C7") +
  xlab(expression("Root mean square error (RMSE)")) +
  scale_y_continuous(position = "right") +
  theme_bw() +
  theme(
    legend.position = "left",
    panel.grid.minor = element_blank()
  ) +
  ylab("")


part1_a<-ggarrange(
  R2_decision,
  RMSE_decision,
  ncol = 2,
  legend = "bottom",
  common.legend = FALSE
) %>%
  annotate_figure(
    top = text_grob(
      "Explained Variance and Error Distribution of the Integrated Model",
      face = "bold",
      size = 14
    )
  )
# cowplot::plot_grid(isoforcePLot,part1_a,labels = c("A", "B"),rel_heights =   c(2.3,0.6),ncol = 1)
# cow(isoforcePLot,part1_a,ncol = 1)
part1_a
ggsave(file="../figures/Supplement_Model_performance2.png",device = "png",width = 9,height = 7)

```

```{r}
eck <- "+proj=eck4"

world <- ne_countries(scale = "medium", returnclass = "sf")  %>%
  st_wrap_dateline(options = c("WRAPDATELINE=YES","DATELINEOFFSET=180"), quiet = TRUE) %>%
  st_transform(crs = eck)

testdata_check<- readRDS("../output/03_2511_testdata")



checkthelist<-testdata_check$test_subset %>% group_by(Country.Code) %>% 
  reframe(RMSE_mean=mean(RMSE),
          RMSE_var=var(RMSE),
          RMSE_n=length(RMSE),
          Rsquared_mean=mean(Rsquared),
          Rsquared_var=var(Rsquared),
         Rsquared_n=length(Rsquared))

groups_dim<-dimred_250 %>%filter(year%in%2019) %>%   select(`Income Group`,`Country Code`,RegionSoviet,iso1) %>%  unique

testdata_check$test_subset

testframe<-inner_join(checkthelist,groups_dim,by=c("Country.Code"="Country Code"))
# 
testframe %>%  ggplot(aes(iso1,RMSE_mean,color=`Income Group`))+geom_point()
```
```{r}
D_error <- testframe %>%
  filter(!is.na(`Income Group`)) %>%
  ggplot(aes(x = `Income Group`, y = `RMSE_mean`, fill = `Income Group`)) +  # Change color to fill
  geom_boxplot(color="black") +
  default_fill_palette+  # Adjust color palette
  default_color_palette_plotA+
  theme(legend.position = "asd", axis.line = element_line(linetype = "dotted")) +
 ylab("Excess Mortality in %") +
  xlab("Income Group") +
  geom_jitter(aes(color = `Income Group`),alpha = 0.6) 
D_error
```


```{r}
world_data <- merge(world,testframe , by.x = "iso_a3_eh", by.y = "Country.Code", all.x = TRUE) 



worldmap_excessMort <- ggplot() +
  geom_sf(data = world_data, aes(fill = RegionsSoviet), lwd = 0.1, size = 6) +
  scale_fill_viridis(begin = 0.0,end=1,
    option = "B",
    name = "Excess Mortality  ")
worldmap_excessMort

 
```



# Q1. Revier request: Estimation vs. Real values excess mortality



```{r}
features_integrated_model <-  c("Vaccinations","COVID-19 cases","Oxford Stringency Index",
                                "iso1","iso2","iso3","iso4","iso5","iso6","iso7","iso8",
                                "iso9","iso10","Obesity (%)","Diabetes (%)","Comorbity",
                                "Urban area (%)")

features_pandemic_indicators <-  c("Vaccinations","COVID-19 cases","Oxford Stringency Index")

features_cNFCs <-  c("iso1","iso2","iso3","iso4","iso5",
                     "iso6","iso7","iso8","iso9","iso10")

features_Epidemiological_features <-  c("Obesity (%)","Diabetes (%)","Comorbity","Urban area (%)")

feature_sets <- list(
  Integrated              = features_integrated_model,
  Pandemic_indicators     = features_pandemic_indicators,
  cNFCs                   = features_cNFCs,
  Epidemiological_features = features_Epidemiological_features
)
library(dplyr)
library(purrr)
library(ggplot2)

y_var <- "p_value_mort"

fit_xgb_for_features <- function(feats, feat_name) {
  # Build design matrix
  dataX <- as.matrix(Xgboostinput_iso %>% dplyr::select(all_of(feats)))
  label_vec <- as.matrix(Xgboostinput_iso[[y_var]])
  
  # Same param_list you defined above
  param_list <- list(
    objective = "reg:squarederror",
    nrounds = c(100,50),
    max_depth = c(3, 2),
    eta = c(0.05),
    gamma = c(0, 5),
    colsample_bytree = c(0.9, 1),
    subsample = c(0.7, 0.8, 0.9),
    min_child_weight = c(5, 10)
  )
  
  mod <- xgboost::xgboost(
    data = dataX,
    label = label_vec,
    params = param_list,
    nrounds = 100,
    verbose = FALSE,
    nthread = 30,
    early_stopping_rounds = 100
  )
  
  predicted <- predict(mod, newdata = dataX)
  actual <- Xgboostinput_iso[[y_var]]
  
  tibble(
    Country    = Xgboostinput_iso$`Country.Code`,
    Actual     = actual,
    Predicted  = predicted,
    FeatureSet = feat_name
  )
}
# Run models for all sets
comparison_all <- map2_dfr(
  feature_sets,
  names(feature_sets),
  ~ fit_xgb_for_features(.x, .y)
)
comparison_all$FeatureSet <- factor(
  comparison_all$FeatureSet,
  levels = c(
    "Pandemic_indicators",
    "Epidemiological_features",
    "cNFCs",
    "Integrated"
  )
)
# Join with income group / GDP info
# make sure order is fixed (if not already done)
comparison_all$FeatureSet <- factor(
  comparison_all$FeatureSet,
  levels = c(
    "Pandemic_indicators",
    "Epidemiological_features",
    "cNFCs",
    "Integrated"
  )
)
```


```{r}
element_a_new <- ggplot(comparison_all, aes(x = Actual, y = Predicted, colour = FeatureSet)) +
  geom_point(alpha = 0.6, size = 2) +
  geom_smooth(method = "lm", se = FALSE, size = 1) +
  geom_abline(intercept = 0, slope = 1, linetype = "dashed") +
  theme_minimal() +
  scale_color_manual(
    values = c(
      "Pandemic_indicators"      = "#D55E00",
      "Epidemiological_features" = "#F0E442",
      "cNFCs"                    = "#009E73",
      "Integrated"               = "#56B4E9"
    ),
    labels = c(
      "Pand.\nind.",    # Pandemic indicators
      "Epid.\nfeat.",   # Epidemiological features
      "cNFC",           # stays as is
      "Int.\nmodel"     # Integrated model
    )
  ) +
  labs(
    x = "Actual values",
    y = "Predicted values",
    colour = "Groups"           # legend title
  ) +
  theme(
    legend.position = "left"    # move legend to the left side
  )

```



### Update 3.1 plot
```{r}

ggarrange(ggarrange(element_a_new,SectionOver+theme(legend.position ="as"),labels = c("a", "b")),SectionUnder+theme(legend.position = "as"),ncol=1,labels = c("", "c"))  %>%  ggsave(file="../figures/04_figures/04_2511_condensed.png",device = "png",width = 9,height = 7)



```



# Q2. Reviewer question: What difference does it make if we include the infections fatality ratio 
## run XGBOOST iso10

```{r}

modeliso10<-readRDS(iso10_model_path)
topruns<-modeliso10$Xgboost_summary %>% arrange(desc(Rsquared))
xgboostresults<-modeliso10$Xgboost_summary
```


```{r}
Xgboostinput_complete_iso <-readRDS(Xgboostinput_complete_iso_path)
# Xgboostinput_complete_iso %>% fwrite("../output/04_final_dataset.csv")
Xgboostinput_iso<-Xgboostinput_complete_iso%>% rename("Country.Code"=`Country Code`,
                                                     "Diabetes (%)" =`Diabetetes (%)`,
                                                     "COVID-19 cases"=`Covid cases per million`,
                                                     "Comorbity"=`Comorbidity index` )

Label<-"p_value_mort"
features<-Xgboostinput_iso %>% select(-`Country.Code`,-p_value_mort) %>%  colnames()
features_iso<-Xgboostinput_iso %>% select(-`Country.Code`,-p_value_mort) %>%  colnames()

features_iso<-features_iso
# Xgboostinput_complete_iso %>% write_csv("../output/thisisisoxgboost.csv")
```


```{r}
incomeGroupDeaths<-merge(Xgboostinput_complete_iso,dimred_250 %>% select(`Country Code`,`Income Group`,RegionSoviet) %>%  unique,by="Country Code")
anova_result<-aov(p_value_mort~`Income Group`,incomeGroupDeaths) 
qqnorm(residuals(anova_result))
qqline(residuals(anova_result))
```



```{r}
df_scaled<-Xgboostinput_iso[,-1] %>% scale() %>% data.frame()
df_scaled_withlabels<-cbind(Xgboostinput_iso[,1],Xgboostinput_iso[,-1])
colnames(df_scaled_withlabels)<-colnames(Xgboostinput_iso) 


Xgboostinput_filtered<-df_scaled_withlabels%>%  filter(!(Country.Code%in% "PER"))
set.seed(topruns$SeedNum[1])
features_iso<-Xgboostinput_filtered %>%  select(-p_value_mort,-Country.Code) %>%  colnames()
features_iso<-features_iso

wdi_imputed_shaptable<-data.frame()
wdi_imputed_bootstrap<-data.frame()


# Split test training
  countries<-Xgboostinput_filtered$Country.Code %>%  unique
  test_proportion <- 0.8
  train_countries <- sample(countries, size = length(countries) * test_proportion)
  train_data <- Xgboostinput_filtered %>% filter(Country.Code%in% train_countries)
  test_data <- Xgboostinput_filtered %>% filter(!(Country.Code%in% train_countries))
  train_data$group <- as.numeric(factor(train_data$Country.Code))
  # Set up trainControl to perform group-wise cross-validation
  folds<-createFolds(train_data$group, k = 5, list = TRUE)
custom_control <- trainControl(
    index = folds,
    method = "repeatedcv",
    number = 10,
    repeats = 5, # Number of repetitions
    verboseIter = TRUE,allowParallel = T
  )
param_grid <- expand.grid(
  nrounds = c(100, 200, 300),  # Number of boosting rounds
  max_depth = c(3, 4, 5),     # Maximum tree depth
  eta = c(0.01, 0.1, 0.3),    # Learning rate
  gamma = c(0, 1, 10),       # L2 regularization
  colsample_bytree = c( 0.9,1),
  subsample = c(0.7, 0.8, 0.9),  # Subsample ratio
  min_child_weight = c(1, 5, 10)  # Minimum child weight
)
Sys.setenv(OPENBLAS_NUM_THREADS = 1)
doMC::registerDoMC(30)


  # Test different parameters
xgb_caret_wdi_raw <- caret::train(
  x = train_data %>% dplyr::select(features_iso),
  y = train_data %>% pull(Label),
  method = "xgbTree",
  tuneGrid = param_grid,
  trControl = custom_control,
  tuneLength = 100,
  verbosity = 0
)



finalmodel_wdi_iso <- caret::train(
    x = train_data %>% dplyr::select(features_iso),
    y = train_data %>% pull(Label),
    method = "xgbTree",
    verbosity = 0,trControl = custom_control,
    tuneGrid = xgb_caret_wdi_raw$bestTune
  )
  
  predictedValue_iso<-postResample(
    predict(finalmodel_wdi_iso, 
            test_data %>% dplyr::select(features_iso)),  
    test_data %>% pull(Label)
  )
predictedValue_iso
train_data_iso<-train_data
shaptable <- xgb.plot.shap.summary(data = as.matrix(train_data %>% select(features_iso)), model = finalmodel_wdi_iso$finalModel, top_n =20)
  
finalmodel_wdi_iso
shaptable
  # shaptable$data %>% group_by(feature) %>% summarise(numberobservations=length(shap_value),Shap_value_mean=mean(abs(shap_value))) %>%  arrange(desc(Shap_value_mean))->shaptable_data
  # 
  # imp<-varImp(finalmodel_wdi_raw)$importance %>%
  #   mutate(feature = rownames(varImp(finalmodel_wdi_raw)$importance))%>%
  #   rename(importance = Overall)%>%
  #   filter(importance > 0)

```



```{r}
Xgboostinput_filtered<-Xgboostinput_filtered %>% data.frame
features<-colnames(Xgboostinput_filtered)[!(Xgboostinput_filtered %>%  colnames() %in%c("Country.Code","p_value_mort"))]


```

## Add IFR

```{r}
ihme_ifr$value_mean %>% summary
ihme_ifr<-fread(file.path(INPUT_DIR,"HME_COVID_19_IES_2019_2021_RATIOS_Y2022M04D08.CSV"))
covid_latest <- owid_data %>% mutate(year = year(date)) %>% filter(year<2023)
numeric_vars <- covid_latest %>%
  dplyr::select(where(is.numeric)) %>%
  colnames()

numeric_vars <- setdiff(numeric_vars, c("year"))  # remove grouping var

# Define cumulative vars (take max per year)
cumulative_vars <- intersect(numeric_vars, c(
  "total_cases", "total_cases_per_million",
  "total_deaths", "total_deaths_per_million",
  "total_tests", "total_tests_per_thousand",
  "total_vaccinations", "people_vaccinated",
  "people_fully_vaccinated", "total_boosters",
  "total_vaccinations_per_hundred", "people_vaccinated_per_hundred",
  "people_fully_vaccinated_per_hundred", "total_boosters_per_hundred"
))

# Flow vars (sum per year)
flow_vars <- setdiff(numeric_vars, cumulative_vars)
```


```{r}
ihme_ifr<-ihme_ifr %>%  filter(measure_name%in%"Cumulative infection-fatality ratio") %>% 
  mutate(year=case_when(grepl("2020-",as.character(date))~"2020",
                        grepl("2021-",as.character(date))~"2021",
                        TRUE~"asd"),
         iso3_code = countrycode(location_name, "country.name", "iso3c", nomatch = location_name))
ihme_ifr_2<-ihme_ifr %>% group_by(iso3_code,year) %>% 
  reframe(IFR_mean=mean(value_mean),
          IFR_median=median(value_mean),
          IFR_sum=sum(value_mean)) %>% mutate(year=as.numeric(year))
datadimred<-dimred_250
datadimred_ifr<-datadimred %>% mutate(year=case_when(year%in%2021~9999,year%in%2020~2021,year%in%2019~2020,TRUE~year)) %>%  filter(year%in%c(2020,2021)) %>% 
  dplyr::select(`Country Code`,year,iso1,iso2,iso3,iso4,iso5) %>% inner_join( . , ihme_ifr_2, by=c("Country Code"="iso3_code","year"="year")) 

covid_years <- covid_latest %>%filter(year%in%c(2020,2021)) %>% 
  group_by(location,year) %>%
  reframe(
    across(all_of(flow_vars), ~ sum(.x, na.rm = TRUE), .names = "sum_{.col}"),
    across(all_of(cumulative_vars), ~ max(.x, na.rm = TRUE), .names = "max_{.col}"),
    .groups = "drop"
  ) %>% suppressWarnings() %>% mutate(iso3_code = countrycode(location, "country.name", "iso3c", nomatch = location))
covid_years_ifr<-covid_years %>% inner_join( . , datadimred_ifr, by=c("iso3_code"="Country Code","year"="year"))
covid_years_ifr_ex<-covid_years_ifr %>% inner_join( . , excessCounrty_annual, by=c("iso3_code"="iso3","year"="year"))
cor(covid_years_ifr_ex$iso1,covid_years_ifr_ex$IFR_median,use = "complete.obs")
cor(covid_years_ifr_ex$sum_new_cases_per_million,covid_years_ifr_ex$IFR_median,use = "complete.obs")
```
```{r}
df_complete <- covid_years_ifr_ex %>%
  dplyr::select(iso1,iso2, IFR_median,sum_new_cases_per_million,p_value_mort) %>%
  na.omit()

# compute distance correlation
dcor(df_complete$iso1, df_complete$IFR_median)
dcor(df_complete$iso2, df_complete$IFR_median)
dcor(df_complete$p_value_mort, df_complete$IFR_median)
dcor(df_complete$sum_new_cases_per_million, df_complete$IFR_median)

```


```{r}
ggplot(covid_years_ifr_ex, aes(IFR_median, p_value_mort,
               color = sum_new_cases_smoothed_per_million)) +
  geom_point(size=2, alpha=0.7) +
  geom_smooth(method="lm", se=TRUE, color="black") +
  scale_color_viridis_c() +
  theme_minimal() +
  labs(
    x = "IFR median",
    y = "p-value (mortality)",
    color = "Cases per million",
    title = "Effect of IFR on mortality p-values"
  )

lm(p_value_mort~sum_new_cases_smoothed_per_million+IFR_median,covid_years_ifr_ex %>% filter(!(iso3_code%in%"PER"))) %>%  summary()
lm(IFR_median~iso1+iso2+iso3+iso4+sum_new_cases_smoothed_per_million+p_value_mort,covid_years_ifr_ex %>% filter(!(iso3_code%in%"PER"))) %>%  summary()
```


```{r}
Xgboostinput_complete_iso <-readRDS(Xgboostinput_complete_iso_path)
# Xgboostinput_complete_iso %>% fwrite("../output/04_final_dataset.csv")
Xgboostinput_iso<-Xgboostinput_complete_iso%>% rename("Country.Code"=`Country Code`,
                                                     "Diabetes (%)" =`Diabetetes (%)`,
                                                     "COVID-19 cases"=`Covid cases per million`,
                                                     "Comorbity"=`Comorbidity index` )

Xgboostinput_iso$year<- rep(c(2020, 2021), each = nrow(Xgboostinput_iso) / 2)
covid_years_ifr_ex$IFR_median<-scale(covid_years_ifr_ex$IFR_median)
mergedFrame<-inner_join(Xgboostinput_iso,covid_years_ifr_ex[,c("iso3_code","year","IFR_median")], by=c("Country.Code"="iso3_code","year"="year"))

```


```{r}
Label<-"p_value_mort"
features<-mergedFrame %>% select(-`Country.Code`,-p_value_mort,-year) %>%  colnames()
features_iso<-mergedFrame %>% select(-`Country.Code`,-p_value_mort,-year) %>%  colnames()

# Xgboostinput_complete_iso %>% write_csv("../output/thisisisoxgboost.csv")
```





```{r}
df_scaled<-mergedFrame[,] %>% select(-Country.Code,-year) %>% scale() %>% data.frame()
df_scaled_withlabels<-bind_cols(mergedFrame %>% select(Country.Code),mergedFrame %>% select(-Country.Code,-year) )
df_scaled_withlabels$IFR_median<-df_scaled_withlabels$IFR_median
df_scaled_withlabels$year<-NULL
```


```{r}

colnames(df_scaled_withlabels)<-str_replace(colnames(df_scaled_withlabels),"iso","cNFC")

Xgboostinput_filtered<-df_scaled_withlabels%>%  filter(!(Country.Code%in% "PER"))
set.seed(topruns$SeedNum[1])
features_iso<-Xgboostinput_filtered %>%  select(-p_value_mort,-Country.Code) %>%  colnames()


wdi_imputed_shaptable<-data.frame()
wdi_imputed_bootstrap<-data.frame()


# Split test training
  countries<-Xgboostinput_filtered$Country.Code %>%  unique
  test_proportion <- 0.8
  train_countries <- sample(countries, size = length(countries) * test_proportion)
  train_data <- Xgboostinput_filtered %>% filter(Country.Code%in% train_countries)
  test_data <- Xgboostinput_filtered %>% filter(!(Country.Code%in% train_countries))
  train_data$group <- as.numeric(factor(train_data$Country.Code))
  # Set up trainControl to perform group-wise cross-validation
  folds<-createFolds(train_data$group, k = 5, list = TRUE)
custom_control <- trainControl(
    index = folds,
    method = "repeatedcv",
    number = 10,
    repeats = 5, # Number of repetitions
    verboseIter = TRUE,allowParallel = T
  )
param_grid <- expand.grid(
  nrounds = c(100, 200, 300),  # Number of boosting rounds
  max_depth = c(3, 4, 5),     # Maximum tree depth
  eta = c(0.01, 0.1, 0.3),    # Learning rate
  gamma = c(0, 1, 10),       # L2 regularization
  colsample_bytree = c( 0.9,1),
  subsample = c(0.7, 0.8, 0.9),  # Subsample ratio
  min_child_weight = c(1, 5, 10)  # Minimum child weight
)
Sys.setenv(OPENBLAS_NUM_THREADS = 1)
doMC::registerDoMC(30)


  # Test different parameters
xgb_caret_wdi_raw <- caret::train(
  x = train_data %>% dplyr::select(features_iso),
  y = train_data %>% pull(Label),
  method = "xgbTree",
  tuneGrid = param_grid,
  trControl = custom_control,
  tuneLength = 100,
  verbosity = 0
)



finalmodel_wdi_iso <- caret::train(
    x = train_data %>% dplyr::select(features_iso),
    y = train_data %>% pull(Label),
    method = "xgbTree",
    verbosity = 0,trControl = custom_control,
    tuneGrid = xgb_caret_wdi_raw$bestTune
  )
  
  predictedValue_iso<-postResample(
    predict(finalmodel_wdi_iso, 
            test_data %>% dplyr::select(features_iso)),  
    test_data %>% pull(Label)
  )
predictedValue_iso
train_data_iso<-train_data
shaptable <- xgb.plot.shap.summary(data = as.matrix(train_data %>% select(features_iso)), model = finalmodel_wdi_iso$finalModel, top_n =20)
  
shaptable %>% ggsave(file = "../figures/04_figures/Supplement_infection_fatality_ratio.png",device = "png")

```


# 3.2 Socioeconomic drivers of pandemic outcomes Figure 4
```{r}
df_scaled<-Xgboostinput_iso[,-1] %>% scale() %>% data.frame()
df_scaled_withlabels<-cbind(Xgboostinput_iso[,1],Xgboostinput_iso[,-1])
colnames(df_scaled_withlabels)<-colnames(Xgboostinput_iso) 

Xgboostinput_filtered<-df_scaled_withlabels%>%  filter(!(Country.Code%in% "PER"))
set.seed(topruns$SeedNum[1])
features_iso<-Xgboostinput_filtered %>%  select(-p_value_mort,-Country.Code) %>%  colnames()


wdi_imputed_shaptable<-data.frame()
wdi_imputed_bootstrap<-data.frame()


set.seed(1212)

# Split test training
countries<-Xgboostinput_filtered$Country.Code %>%  unique
test_proportion <- 1
train_countries <- sample(countries, size = length(countries) * test_proportion)
train_data <- Xgboostinput_filtered %>% filter(Country.Code%in% train_countries)
train_data$group <- as.numeric(factor(train_data$Country.Code))
folds<-createFolds(train_data$group, k = 5, list = TRUE)

# Load the necessary libraries
library(doParallel)
library(caret)

# Register 30 cores for parallel processing
modeliso10$Xgboost_summary->resultsEval
resultsEval$trainR2 %>% mean
set.seed(1212)
# Custom trainControl with repeated cross-validation
custom_control <- trainControl(
  index = folds,
  method = "repeatedcv",
  number = 10,
  repeats = 5,  # Number of repetitions
  verboseIter = TRUE,
  allowParallel = TRUE
)

# Parameter grid for XGBoost
param_grid <- expand.grid(
      nrounds = c(300),  # Number of boosting rounds
      max_depth = c( 3),     # Maximum tree depth
      eta = c(0.01),    # Learning rate
      gamma = c(1),       # L2 regularization
      colsample_bytree = c( 0.8),
      subsample = c(0.8),  # Subsample ratio
      min_child_weight = c(5)
    )

# Train the model using caret with parallel processing
xgb_caret_wdi_raw <- caret::train(
  x = train_data %>% dplyr::select(features_iso),
  y = train_data %>% pull(Label),
  method = "xgbTree",
  tuneGrid = param_grid,
  trControl = custom_control,
  verbosity = 1
)

# Use the best model parameters to train the Final model
finalmodel_wdi_iso <- caret::train(
  x = train_data %>% dplyr::select(features_iso),
  y = train_data %>% pull(Label),
  method = "xgbTree",
  tuneGrid = xgb_caret_wdi_raw$bestTune,
  verbosity = 0
)


# Stop the cluster after training is complete

train_data_iso<-train_data
  shaptable <- xgb.plot.shap.summary(data = as.matrix(train_data %>% dplyr::select(features_iso)), model = finalmodel_wdi_iso$finalModel, top_n =20)

col_pal <- c("grey80",  # Red
                   "#377EB8",  # Blue
                   "#4DAF4A",  # Green
                   "#984EA3",  # Purple
                   "#FF7F00",  # Orange
                   "#FFFF33",  # Yellow
                   "#A65628",  # Brown
                   "#F781BF",  # Pink
                   "firebrick3")  # Gray
# col_pal <- c("#8dd3c7", "#ffffb3", "#bebada", "#fb8072", "#80b1d3", 
#             "#fdb462", "#b3de69", "#fccde5", "#d9d9d9", "#bc80bd", 
#             "#ccebc5", "#ffed6f", "black")


theme_set(
  theme_bw(
    base_size = getOption("book.base_size"),
    base_family = getOption("book.base_family")
  ) %+replace%
    theme(
      panel.grid.minor = element_blank(),
      legend.position = "bottom"
    )
)
options(
  # set default colors in ggplot2 to colorblind-friendly
  # Okabe-Ito and Viridis palettes
  ggplot2.discrete.colour = col_pal,
  ggplot2.discrete.fill = col_pal,
  ggplot2.continuous.colour = "viridis",
  ggplot2.continuous.fill = "viridis",
  # set theme font and size
  book.base_family = "sans",
  book.base_size = 12
)
 shaptable <- xgb.plot.shap.summary(data = as.matrix(train_data %>% select(features_iso)), model = finalmodel_wdi_iso$finalModel, top_n =20)
 

library(ggExtra)
train_data2<-train_data %>%  mutate(Country.Code_label=case_when(
                                                    Country.Code%in%"RUS"~"Russia",
                                                    Country.Code%in%"DEU"~"Germany",
                                                    # 
                                                    Country.Code%in%"USA"~"USA",
                                                    Country.Code%in%"BRA"~"Brazil",
                                                    # 
                                                    # Country.Code%in%"THA"~"Thailand",
                                                    # Country.Code%in%"KOR"~"Korea",
                                                    # 
                                                    # Country.Code%in%"IDN"~"Indonesia",
                                                    # Country.Code%in%"NAM"~"Namibia",
                                                    # 
                                                    # Country.Code%in%"PRY"~"Paraguay",
                                                    # Country.Code%in%"TUN"~"Tunesia",
                                                    # 
                                                    # Country.Code%in%"COD"~"Congo (REP)",
                                                    # Country.Code%in%"DZA"~"Algeria",
                                                    # 
                                                    
                                                    # 
                                                    
                                                    # Country.Code%in%"YEM"~"Yemen",
                                                    TRUE~""
                                                    ))%>%
  mutate(dot_size = ifelse(Country.Code_label == "","a", "b"))%>%
  mutate(alpha = ifelse(Country.Code_label == "","a", "b")) %>% inner_join( . ,excessCounrty_annual[,c("p_value_mort","year")] ) 

p2 <- ggMarginal((cbind(train_data2,shaptable$data %>%  filter(feature%in%c("iso2") ))%>% 
  ggplot()+geom_point(aes(x=feature_value,y=shap_value,color=Country.Code_label,alpha=as.factor(alpha),size=as.factor(dot_size)))+scale_size_manual(values=c(1,3))+ylab("SHAP values") +scale_alpha_manual(values = c(0.4,1))+scale_size_manual(values=c(2,5))+ geom_smooth(aes(x=feature_value,y=shap_value))+theme(legend.position = "none") +xlab("cNFC 2")), type="histogram")
p3 <- ggMarginal((cbind(train_data2,shaptable$data %>%  filter(feature%in%c("iso3") ))%>% 
  ggplot()+geom_point(aes(x=feature_value,y=shap_value,color=Country.Code_label,size=dot_size,alpha=as.factor(alpha)))+scale_alpha_manual(values = c(0.4,1))+scale_size_manual(values=c(2,5))+ylab("SHAP values") + geom_smooth(aes(x=feature_value,y=shap_value))+theme(legend.position = "none")  +xlab("cNFC 3")), type="histogram")
p8 <- ggMarginal((cbind(train_data2,shaptable$data %>%  filter(feature%in%c("iso8") ))%>% 
  ggplot()+geom_point(aes(x=feature_value,y=shap_value,color=Country.Code_label,size=dot_size,alpha=as.factor(alpha)))+scale_alpha_manual(values = c(0.4,1))+scale_size_manual(values=c(2,5))+ylab("SHAP values") + geom_smooth(aes(x=feature_value,y=shap_value))+theme(legend.position = "none") +xlab("cNFC 8 ")), type="histogram")
p4 <- ggMarginal((cbind(train_data2,shaptable$data %>%  filter(feature%in%c("iso4") ))%>% 
  ggplot()+geom_point(aes(x=feature_value,y=shap_value,color=Country.Code_label,size=dot_size,alpha=as.factor(alpha)))+scale_alpha_manual(values = c(0.4,1))+scale_size_manual(values=c(2,5))+ylab("SHAP values") + geom_smooth(aes(x=feature_value,y=shap_value))+theme(legend.position = "none")  +xlab("cNFC 4")), type="histogram")
p1 <- ggMarginal((cbind(train_data2,shaptable$data %>%  filter(feature%in%c("iso1") ))%>% 
  ggplot()+geom_point(aes(x=feature_value,y=shap_value,color=Country.Code_label,size=dot_size,alpha=as.factor(alpha)))+scale_alpha_manual(values = c(0.4,1))+scale_size_manual(values=c(2,5))+ylab("SHAP values") + geom_smooth(aes(x=feature_value,y=shap_value))+theme(legend.position = "none") +xlab("cNFC 1")), type="histogram")
pCases <- ggMarginal((cbind(train_data2,shaptable$data %>%  filter(feature%in%c("COVID-19 cases") ))%>% 
ggplot()+geom_point(aes(x=`COVID-19 cases`,y=shap_value,color=Country.Code_label,size=dot_size,alpha=as.factor(alpha)))+scale_alpha_manual(values = c(0.4,1))+scale_size_manual(values=c(2,5))+ylab("SHAP values") +geom_smooth(aes(x=`COVID-19 cases`,y=shap_value))+theme(legend.position = "none") +xlab("COVID-19 cases")), type="histogram")





Plots<-ggarrange(pCases,p1,p2,p3,p4,p8,ncol = 2,nrow = 3,labels = c("a","b","c","d","e","f"))

# Plots %>% ggsave(file="../figures/04_figures/04_2412_discussion_new.png",width = 12,height = 8)

#legndi %>% ggsave(file="../figures/04_figures/04_legend_discussion.png")


legendA <- ggMarginal((cbind(train_data2,shaptable$data %>%  filter(feature%in%c("COVID-19 cases") )) %>% mutate(Countries=Country.Code_label,year=as.factor(year)) %>%  filter(!(Countries%in%"")) %>%  
ggplot()+geom_point(aes(x=`COVID-19 cases`,y=shap_value,color=Countries))+scale_size_manual(values=c(0.5,3))+ylab("SHAP values") +theme(legend.position = "bottom") +xlab("COVID-19 cases")+scale_color_manual(values = col_pal[c(2:length(col_pal))])), type="histogram")
legendA
legendB <- ggMarginal((cbind(train_data2,shaptable$data %>%  filter(feature%in%c("COVID-19 cases") )) %>% mutate(Countries=Country.Code_label,Year=as.factor(year)) %>%  
ggplot()+geom_point(aes(x=`COVID-19 cases`,y=shap_value,shape = Year))+ylab("SHAP values") +theme(legend.position = "right") +xlab("COVID-19 cases")), type="histogram")
legend1 = get_plot_component(legendA, 'guide-box-bottom', return_all = TRUE)
legend2 = get_plot_component(legendB, 'guide-box-bootm', return_all = TRUE)
legend=plot_grid(legend1,legend2,align = "v", axis = "tb")
legend_centered = plot_grid(
  NULL, legend1, NULL, 
  ncol = 3, 
  rel_widths = c(1, 0.5, 1) # Adjust to center the legend
)
legend_centered2 = plot_grid(
  NULL, legend2, NULL, 
  ncol = 3, 
  rel_widths = c(1, 0.5, 1) # Adjust to center the legend
)
legend #%>% ggsave(file="../figures/04_figures/04_legend_discussion.png",width = 12,height = 8)
final_plot = plot_grid(
  Plots, 
  legend_centered, 
  ncol = 1,          # Arrange in one column (vertical stacking)
  rel_heights = c(1, 0.1),  # Adjust heights, give more space to panelA
  align = "v"        # Align vertically
)

final_plot # %>%  ggsave(file="../figures/04_figures/04_2506_discussion_new.png",width = 8,height = 8)

# final_plot %>%  ggsave(file="../figures/Lancet_submission/Results3.pdf",width = 8,height = 8)
```
### Supplement versions
```{r}


library(ggExtra)
regionsMapper<-dimred_250 %>%  select(`Country Code`,`RegionSoviet`) %>%  unique
train_data2<-train_data%>% inner_join( . ,excessCounrty_annual[,c("p_value_mort","year")] ) %>% inner_join( . ,regionsMapper,by=c(Country.Code="Country Code" ))   %>%
    mutate(Country.Code_label=RegionSoviet) %>% 
  mutate(dot_size = ifelse(Country.Code_label == "","a", "b"),alpha = "1")
p2 <- ggMarginal((cbind(train_data2,shaptable$data %>%  filter(feature%in%c("iso2") ))%>% 
  ggplot()+geom_point(aes(x=feature_value,y=shap_value,color=Country.Code_label,alpha=as.factor(alpha),size=as.factor(dot_size)))+scale_size_manual(values=c(1,3))+ylab("SHAP values") +scale_alpha_manual(values = c(1))+scale_size_manual(values=c(2,5))+ geom_smooth(aes(x=feature_value,y=shap_value))+theme(legend.position = "none") +xlab("cNFC 2")), type="histogram")
p3 <- ggMarginal((cbind(train_data2,shaptable$data %>%  filter(feature%in%c("iso3") ))%>% 
  ggplot()+geom_point(aes(x=feature_value,y=shap_value,color=Country.Code_label,size=dot_size,alpha=as.factor(alpha)))+scale_alpha_manual(values = c(1))+scale_size_manual(values=c(2,5))+ylab("SHAP values") + geom_smooth(aes(x=feature_value,y=shap_value))+theme(legend.position = "none")  +xlab("cNFC 3")), type="histogram")
p8 <- ggMarginal((cbind(train_data2,shaptable$data %>%  filter(feature%in%c("iso8") ))%>% 
  ggplot()+geom_point(aes(x=feature_value,y=shap_value,color=Country.Code_label,size=dot_size,alpha=as.factor(alpha)))+scale_alpha_manual(values = c(1))+scale_size_manual(values=c(2,5))+ylab("SHAP values") + geom_smooth(aes(x=feature_value,y=shap_value))+theme(legend.position = "none") +xlab("cNFC 8 ")), type="histogram")
p4 <- ggMarginal((cbind(train_data2,shaptable$data %>%  filter(feature%in%c("iso4") ))%>% 
  ggplot()+geom_point(aes(x=feature_value,y=shap_value,color=Country.Code_label,size=dot_size,alpha=as.factor(alpha)))+scale_alpha_manual(values = c(1))+scale_size_manual(values=c(2,5))+ylab("SHAP values") + geom_smooth(aes(x=feature_value,y=shap_value))+theme(legend.position = "none")  +xlab("cNFC 4")), type="histogram")
p1 <- ggMarginal((cbind(train_data2,shaptable$data %>%  filter(feature%in%c("iso1") ))%>% 
  ggplot()+geom_point(aes(x=feature_value,y=shap_value,color=Country.Code_label,size=dot_size,alpha=as.factor(alpha)))+scale_alpha_manual(values = c(1))+scale_size_manual(values=c(2,5))+ylab("SHAP values") + geom_smooth(aes(x=feature_value,y=shap_value))+theme(legend.position = "none") +xlab("cNFC 1")), type="histogram")
pCases <- ggMarginal((cbind(train_data2,shaptable$data %>%  filter(feature%in%c("COVID-19 cases") ))%>% 
ggplot()+geom_point(aes(x=`COVID-19 cases`,y=shap_value,color=Country.Code_label,size=dot_size,alpha=as.factor(alpha)))+scale_alpha_manual(values = c(1))+scale_size_manual(values=c(2,5))+ylab("SHAP values") +geom_smooth(aes(x=`COVID-19 cases`,y=shap_value))+theme(legend.position = "none") +xlab("COVID-19 cases")), type="histogram")

pCases



Plots<-ggarrange(pCases,p1,p2,p3,p4,p8,ncol = 2,nrow = 3,labels = c("a","b","c","d","e","f"))

# Plots %>% ggsave(file="../figures/04_figures/04_2412_discussion_new.png",width = 12,height = 8)

#legndi %>% ggsave(file="../figures/04_figures/04_legend_discussion.png")


legendA <- ggMarginal((cbind(train_data2,shaptable$data %>%  filter(feature%in%c("COVID-19 cases") ))%>%
ggplot()+geom_point(aes(x=`COVID-19 cases`,y=shap_value,color=Country.Code_label,size=dot_size,alpha=as.factor(alpha)))+scale_alpha_manual(values = c(0.4,1))+scale_size_manual(values=c(2,5))+ylab("SHAP values") +geom_smooth(aes(x=`COVID-19 cases`,y=shap_value))+xlab("COVID-19 cases")), type="histogram")



legend1 = get_plot_component(legendA, 'guide-box-bottom', return_all = TRUE)
legend2 = get_plot_component(legendB, 'guide-box-bootm', return_all = TRUE)
legend=plot_grid(legend1,legend2,align = "v", axis = "tb")
legend_centered = plot_grid(
  NULL, legend1, NULL, 
  ncol = 3, 
  rel_widths = c(1, 0.5, 1) # Adjust to center the legend
)
legend_centered

legend #%>% ggsave(file="../figures/04_figures/04_legend_discussion.png",width = 12,height = 8)
final_plot = plot_grid(
  Plots, 
  legend_centered, 
  ncol = 1,          # Arrange in one column (vertical stacking)
  rel_heights = c(1, 0.1),  # Adjust heights, give more space to panelA
  align = "v"        # Align vertically
)
final_plot
 final_plot  %>%  ggsave(file="../figures/04_figures/04_2506_discussion_supplement.png",width = 8,height = 8)
```


```{r}
summaryofindicator <- function(indicator,year=2020) {
  dimred_250 %>% 
    filter(year %in% c(2020)) %>% 
    select(`Country Code`, indicator) %>% 
    group_by(`Country Code`) %>% 
    reframe(
      meanoftis = mean(get(indicator), na.rm = TRUE),
      # sd = sd(get(indicator), na.rm = TRUE)
    ) %>% cbind( . ,name=wdi_series[wdi_series$`Indicator Code`%in% paste(indicator),`Indicator Name`])
}

# Example usage:


```

```{r}
owid_biannual<-owid_data %>%  
  mutate(year=year(date))  %>% 
  filter(year%in%c(2020,2021)) %>% 
  data.table() %>% select(-date) %>% 
  filter(!is.na(continent)& continent!="") %>%   
  group_by(iso_code,location) %>% 
  reframe(
            Stringency_index=mean(stringency_index,na.rm=T),
            Stringency_index_var=var(stringency_index,na.rm=T),
            # Average_hosp_admissions=sum(weekly_hosp_admissions,na.rm=T)/population,
            # Average_ICU_admissions=sum(weekly_icu_admissions,na.rm=T)/population,
            Covid_Cases_per_million=sum(new_cases_per_million,na.rm=T),
            Owid_death_per_million =sum(new_deaths_per_million,na.rm=T),
            Owid_Exmort =mean(excess_mortality,na.rm=T),
            Vaccinations=sum(total_vaccinations_per_hundred,na.rm=T),
            Available_Hospital_beds=mean(hospital_beds_per_thousand,na.rm=T),
            Median_age=mean(median_age,na.rm=T),
            aged_65_older=mean(aged_65_older,na.rm=T),
            sum_tests_per_thousend=sum(total_tests_per_thousand,na.rm=T),
            human_development_index=mean(human_development_index,na.rm=T),
            handwashing_facilities=mean(handwashing_facilities,na.rm=T),
            population_density=mean(population_density,na.rm=T),
            diabetes_prevalence=mean(diabetes_prevalence,na.rm=T),
            continent=continent,
            cardiovasc_death_rate=mean(cardiovasc_death_rate,na.rm=T)
            
            )%>%  arrange(iso_code) %>%  unique %>% mutate(identifier=paste0(iso_code))
```



```{r}
 
options(scipen=999)
errorsummary<-lm(`COVID-19 cases`+iso1+iso4~p_value_mort,Xgboostinput_iso)
cbind(error=abs(errorsummary$residuals),Xgboostinput_iso) %>% arrange((error))
```

### Russia vs. Germany
```{r}
#CO2 emissions (kg per 2015 US$ of GDP)

summaryofindicator("EN.ATM.CO2E.KD.GD") %>%  filter(`Country Code`%in% c("DEU","RUS"))
# Labor force participation rate for ages 15-24, female (%) (modeled ILO estimate)

summaryofindicator("SL.TLF.ACTI.1524.ZS") %>%  filter(`Country Code`%in% c("DEU","RUS"))

#Employment to population ratio, ages 15-24, total (%) (modeled ILO estimate)


summaryofindicator("SL.EMP.1524.SP.ZS") %>%  filter(`Country Code`%in% c("DEU","RUS"))

summaryofindicator("SL.UEM.1524.ZS") %>%  filter(`Country Code`%in% c("DEU","RUS"))

summaryofindicator("SP.POP.65UP.TO.ZS") %>%  filter(`Country Code`%in% c("DEU","RUS"))




```
```{r}
require(dplyr)
Xgboostinput_iso[Xgboostinput_iso$Country.Code%in%c("RUS","DEU")] %>% 
  dplyr::select(Country.Code,`Covid cases per million`,p_value_mort)

```


```{r}

dimred_250 %>% filter(`Country Code`%in%c("ZAF","USA","BRA","BOL"))  %>% ggplot(aes(y=`SP.POP.2529.FE.5Y`,color=`Country Code`)) +
  geom_boxplot()
```




### Brazil vs. USA
```{r}
## Iso1 
#Vulnerable employment, female (% of female employment) (modeled ILO estimate)
summaryofindicator("SL.EMP.VULN.FE.ZS") %>%  filter(`Country Code`%in% c("BRA","USA"))
#Current health expenditure per capita, PPP (current international $)
summaryofindicator("SH.XPD.CHEX.PP.CD") %>%  filter(`Country Code`%in% c("BRA","USA"))

## iso 8 differnces

summaryofindicator("SL.UEM.TOTL.FE.NE.ZS") %>%  filter(`Country Code`%in% c("BRA","USA"))

## iso 3differnces
#Net ODA received (% of gross capital formation)
summaryofindicator("DT.ODA.ODAT.GI.ZS",2020) %>%  filter(`Country Code`%in% c("BRA","USA"))
#Population in urban agglomerations of more than 1 million (% of total population)
summaryofindicator("EN.URB.MCTY.TL.ZS",2020) %>%  filter(`Country Code`%in% c("BRA","USA"))
#Compensation of employees (% of expense)

summaryofindicator("GC.XPN.COMP.ZS",2020) %>%  filter(`Country Code`%in% c("BRA","USA"))

summaryofindicator("SL.TLF.CACT.FM.ZS",2020) %>%  filter(`Country Code`%in% c("BRA","USA"))

```



```{r}
summaryofindicator_showcase <- function(indicator) {
  library(dplyr)
  library(tidyr)
  this<-dataINf[dataINf$`Indicator Code`%in% indicator,]
  maxcol<-colnames(this[, grepl("iso",colnames(this)),with = F])[max.col(this[, grepl("iso",colnames(this)),with = F])]

  summary <- dimred_250 %>% 
    filter(year %in% 2000:2020) %>% 
    filter(`Country Code` %in% c("DEU", "USA", "BRA", "RUS")) %>%
    group_by(`Country Code`) %>% 
    summarise(
      med = median(.data[[indicator]], na.rm = TRUE),
      iqr = IQR(.data[[indicator]], na.rm = TRUE),
      .groups = "drop"
    ) %>%
    mutate(value = paste0(round(med, 2), " (", round(iqr, 2), ")")) %>%
    mutate(value = as.numeric(round(med, 2))) %>%
    select(`Country Code`, value) %>%
    pivot_wider(names_from = `Country Code`, values_from = value) %>%
    mutate(indicator = wdi_series[wdi_series$`Indicator Code` == indicator, `Indicator Name`],Isocor=maxcol) %>%
    select(indicator,Isocor,DEU, RUS, USA, BRA)

  return(summary)
}

mytable<-rbind(iso1_example_Lifeexptancy<-summaryofindicator_showcase("SP.DYN.LE00.IN"),
               
iso1_example_labor<-summaryofindicator_showcase("NY.GDP.PCAP.PP.KD"),
iso1_example_healthcare<-summaryofindicator_showcase("SH.MED.PHYS.ZS"),
iso2_example_labor<-summaryofindicator_showcase("SL.TLF.ACTI.1524.ZS"),



iso2_example_healthexpenditure<-summaryofindicator_showcase("SH.XPD.CHEX.PC.CD"),
iso2_example_carobin<-summaryofindicator_showcase("NY.ADJ.DCO2.GN.ZS"),
iso3_example_deathrate<-summaryofindicator_showcase("SP.DYN.CDRT.IN"),
iso3_example_deathrate<-summaryofindicator_showcase("SL.TLF.TOTL.FE.ZS"),
iso4_example_deathrate<-summaryofindicator_showcase("VC.IHR.PSRC.P5"),
iso8_example_deathrate<-summaryofindicator_showcase("ST.INT.RCPT.XP.ZS"),
iso8_example_deathrate<-summaryofindicator_showcase("SH.HIV.INCD.TL.P3"))

log10(iso1_example_labor$DEU)
#mytable %>% fwrite("../output/2506_Indicator_group_table.csv")
```





### South Korea and Thailand
```{r}

```

```{r}
table2019_pc<-dimred_250 %>% filter(year%in%2019)
table2020_pc<-dimred_250 %>% filter(year%in%2020) 
table2019_pc2<-table2019_pc %>% mutate(identifier=paste0(`Country Code`,"2020"))
table2020_pc2<-table2020_pc %>% mutate(identifier=paste0(`Country Code`,"2021"))
WDI_pc_2020_2021<-rbind(table2019_pc2,table2020_pc2) %>%  filter(!is.na(iso1))
Xgboostinput<- dplyr::inner_join(WDI_pc_2020_2021, excessCounrty_annual2, by = "identifier")

```
```{r}
cor(Xgboostinput$iso2,Xgboostinput$SH.XPD.CHEX.PC.CD,use = "complete.obs") 
```
